% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_realted.R
\name{annotate_retained_introns}
\alias{annotate_retained_introns}
\title{Annotate introns.}
\usage{
annotate_retained_introns(
  x,
  query_introns,
  add_meta_data = FALSE,
  bs_genome_object = NULL,
  add_prefix_chr = TRUE,
  remove_prefix_chr = FALSE
)
}
\arguments{
\item{x}{an object of class 'parcutils_ir'.}

\item{query_introns}{a character vector of intron_ids to be queried.}

\item{add_meta_data}{logical, default FALSE, whether to add intron meta-data (seq, GC and lenght) to the output.}

\item{bs_genome_object}{an object of class bs_genome. This will be used to retrieve meta-data.
Default  NULL. Required when add_meta_data = TRUE.}

\item{add_prefix_chr}{logical, default TRUE, denoting whether to add prefix 'chr' in the column seqnames.}

\item{remove_prefix_chr}{logical, default FALSE, denoting whether to remove prefix 'chr' from the column seqnames.}
}
\value{
a data frame
}
\description{
Annotate introns.
}
\examples{
example_dir <- system.file("extdata" ,"ir_data" , package="parcutils")
example_files <- fs::dir_ls(example_dir, glob = "*-IR-nondir*.txt")
names(example_files) <- stringr::str_replace(example_files , pattern = ".*/", replacement = "") \%>\%
stringr::str_replace(pattern = "-IR-nondir.txt", replacement = "")
x <- read_irfinderS_output(files = example_files,  add_prefix_chr = F)
ir_sample_info <- names(example_files) \%>\%
tibble::tibble(samples = . , groups = rep(c("control", "treatment"), each = 2) )
run_deseq_analysis_ir(x = x, sample_info = ir_sample_info,
min_intron_depth = 2, regul_based_upon =1, cutoff_pval = 0.01)

q_intron_id = c("intron_223342","intron_223728","intron_225394","intron_228167","intron_228226")
annotate_retained_introns(x, query_introns = q_intron_id)
}
