% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{categorize_diff_genes}
\alias{categorize_diff_genes}
\title{Categorize diff genes based on log2FC, p-value and p-adj value.}
\usage{
categorize_diff_genes(
  dsr_tibble,
  log2fc_cutoff = 1,
  pval_cutoff = 0.05,
  padj_cutoff = 0.01,
  add_column_regul = TRUE,
  regul_based_upon = 1
)
}
\arguments{
\item{dsr_tibble}{a data frame obtained from DESeqResult object or an object of DESeqResult.}

\item{log2fc_cutoff, pval_cutoff, padj_cutoff}{a numeric value, default \verb{(log2fc_cutoff = 1, pval_cutoff =  0.05, padj_cutoff = 0.01)}
denoting cutoffs. These criteria will be used to decide significance  \verb{(NS, p-value, log2FC, p-value&log2FC)} and
type of regulation \verb{(Up, Down & other)} of diff genes.
See details.}

\item{add_column_regul}{logical, default \code{TRUE}, indicating whether to add column  \code{add_column_regul} or not. Added column will contain the values
Up', 'Down' or 'other'.}

\item{regul_based_upon}{one of the numeric choices  1, 2, or 3.
\subsection{if 1 ...}{
\itemize{
\item Up : log2fc >= log2fc_cutoff & pvalue <= pvalue_cutoff
\item Down : log2fc  <= (-1) * log2fc_cutoff & pvalue <= pvalue_cutoff
\item Other : remaining genes
}
}

\subsection{if 2 ...}{
\itemize{
\item Up : log2fc >= log2fc_cutoff & padj <= padj_cutoff
\item Down : log2fc  <= (-1) * log2fc_cutoff & padj <= padj_cutoff
\item Other : remaining genes
}
}

\subsection{if 3 ...}{
\itemize{
\item Up : log2fc >= log2fc_cutoff & pvalue <= pvalue_cutoff & padj <= padj_cutoff
\item Down : log2fc  <= (-1) * log2fc_cutoff pvalue <= pvalue_cutoff & padj <= padj_cutoff
\item Other : remaining genes
}
}}
}
\value{
a data frame
}
\description{
// TO DO
}
\details{
// TO DO. --> explain columns 'signif' and 'type' in the returned data frame.
}
\examples{
\dontrun{

}
}
