% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_realted.R
\name{.assign_class_irFinderSdata}
\alias{.assign_class_irFinderSdata}
\title{Assign class irFinderSdata}
\usage{
.assign_class_irFinderSdata(x)
}
\arguments{
\item{x}{a list or dataframe to which class irFinderSdata to assign.}
}
\value{
an object of class irFinderSdata
}
\description{
This function does all mandatory checks before it assigns class irFinderSdata
}
\keyword{internal}
