% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_realted.R
\name{.assign_intron_identifier}
\alias{.assign_intron_identifier}
\title{Assign unique identifier to each intron.}
\usage{
.assign_intron_identifier(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Assign unique identifier to each intron.
}
\keyword{internal}
