% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{.categorize_diff_ase}
\alias{.categorize_diff_ase}
\title{Categorize diff ASE in to \code{up} and \code{down} based on log2FC, p-value and p-adj values.}
\usage{
.categorize_diff_ase(
  ase_tibble,
  log2fc_cutoff = 1,
  pval_cutoff = 0.05,
  padj_cutoff = 0.01,
  add_column_regul = TRUE,
  regul_based_upon = 1
)
}
\arguments{
\item{ase_tibble}{a dataframe obtained from \code{SpliceWiz::ASE_*} function.}

\item{log2fc_cutoff, pval_cutoff, padj_cutoff}{a numeric value, default \verb{(log2fc_cutoff = 1, pval_cutoff =  0.05, padj_cutoff = 0.01)}
denoting cutoffs. These criteria will be used to decide significance  \verb{(NS, p-value, log2FC, p-value&log2FC)} and
type of regulation \verb{(Up, Down & other)} of diff genes.
See details.}

\item{add_column_regul}{logical, default \code{TRUE}, indicating whether to add column  \code{add_column_regul} or not. Added column will contain the values
Up', 'Down' or 'other'.}

\item{regul_based_upon}{one of the numeric choices  1, 2, or 3.
\subsection{if 1 ...}{
\itemize{
\item Up : log2fc >= log2fc_cutoff & pvalue <= pvalue_cutoff
\item Down : log2fc  <= (-1) * log2fc_cutoff & pvalue <= pvalue_cutoff
\item Other : remaining genes
}
}

\subsection{if 2 ...}{
\itemize{
\item Up : log2fc >= log2fc_cutoff & padj <= padj_cutoff
\item Down : log2fc  <= (-1) * log2fc_cutoff & padj <= padj_cutoff
\item Other : remaining genes
}
}

\subsection{if 3 ...}{
\itemize{
\item Up : log2fc >= log2fc_cutoff & pvalue <= pvalue_cutoff & padj <= padj_cutoff
\item Down : log2fc  <= (-1) * log2fc_cutoff pvalue <= pvalue_cutoff & padj <= padj_cutoff
\item Other : remaining genes
}
}}
}
\value{
a data frame
}
\description{
Usually fold change 1.5 (log2FC ~0.6) and statistical significance defined by p and/or p.adj values
are considered to mark gene (ASE) as differential expressed. However, these cutoffs change based on data quality.
This function allows changing these cutoffs and mark genes (ASE) differently expressed.
While using this package, most of the time user do not need to call this function explicitly as it is internally called by
\code{\link[=run_ase_diff_analysis]{run_ase_diff_analysis()}}.
}
\details{
// TO DO. --> explain columns 'signif' and 'type' in the returned data frame.
}
\keyword{internal}
