% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.check_mendate_columns_for_irFinderSdata}
\alias{.check_mendate_columns_for_irFinderSdata}
\title{Check mandatory columns in a dataframe}
\usage{
.check_mendate_columns_for_irFinderSdata(
  x,
  mandat_columns = c("chr", "start", "end", "name", "null", "strand", "coverage",
    "introndepth", "spliceleft", "spliceright", "spliceexact", "irratio", "warnings")
)
}
\arguments{
\item{x}{dataframe}

\item{mandat_columns}{a character vector denoting mandatory columns in x.}
}
\value{
TRUE or ERROR
}
\description{
This function helps to create mandatory column in a dataframe for the object irFinderSdata.
}
\keyword{internal}
