% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.count_data_convert_NA_to_0}
\alias{.count_data_convert_NA_to_0}
\title{Convert NA to 0 from into count data for the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}}
\usage{
.count_data_convert_NA_to_0(count_data)
}
\arguments{
\item{count_data}{a dataframe}
}
\value{
a dataframe
}
\description{
Convert NA to 0 from into count data for the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}
}
\keyword{internal}
