% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.dsr_to_tibble}
\alias{.dsr_to_tibble}
\title{Convert DESeq result object in to a tibble.}
\usage{
.dsr_to_tibble(x, .col_gene_id = "gene_id")
}
\arguments{
\item{x}{DEseq2 result object. Can be generated by \code{\link[DESeq2:DESeqResults]{DESeq2::DESeqResults()}}.}
}
\value{
a data frame.
}
\description{
Convert DESeq result object in to a tibble.
}
\keyword{internal}
