% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.filter_gff}
\alias{.filter_gff}
\title{Filter rows from a GFF file.}
\usage{
.filter_gff(
  gtf_file,
  feature_type = "gene",
  feature_biotype = "protein_coding",
  var_gene_name = "gene_name",
  var_gene_id = "gene_id",
  var_feature_type = "type",
  var_feature_biotype = "gene_biotype"
)
}
\arguments{
\item{gtf_file}{// TO DO}

\item{feature_type}{// TO DO}

\item{feature_biotype}{// TO DO}

\item{var_gene_name}{// TO DO, to be implement yet.}

\item{var_gene_id}{// TO DO}

\item{var_feature_type}{// TO DO}

\item{var_feature_biotype}{// TO DO}
}
\value{
// TO DO
}
\description{
Filter rows from a GFF file.
}
\keyword{internal}
