% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{.filter_orgdb_by_genetype}
\alias{.filter_orgdb_by_genetype}
\title{Filter selected orgdb columns by gene-type}
\usage{
.filter_orgdb_by_genetype(
  orgdb,
  col_genetype = "GENETYPE",
  feature_type = "protein-coding",
  col_return = "ENTREZID"
)
}
\arguments{
\item{orgdb}{an OrgDB}

\item{col_genetype}{a column name to filter by the value given in \code{feature_type}}

\item{feature_type}{a string denoting value to filter the column mentioned by \code{col_genetype}.}

\item{col_return}{a character string denoting a column to return from an OrgDB.}
}
\value{
a character string of the column mentioned in \code{col_return}.
}
\description{
Filter selected orgdb columns by gene-type
}
\examples{
.filter_orgdb_by_genetype(orgdb = org.Hs.eg.db::org.Hs.eg.db ,col_return = "ENTREZID",feature_type = "snRNA")
.filter_orgdb_by_genetype(orgdb = org.Hs.eg.db::org.Hs.eg.db ,col_return = "ENTREZID",feature_type = "protein-coding")
.filter_orgdb_by_genetype(orgdb = org.Hs.eg.db::org.Hs.eg.db ,col_return = "ENTREZID",feature_type = "awdfd")

}
\keyword{{internal}.}
