% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.filter_rows_from_count_data}
\alias{.filter_rows_from_count_data}
\title{Filter count data using user thresholds for the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}}
\usage{
.filter_rows_from_count_data(
  count_data,
  column_geneid,
  sample_info,
  min_counts,
  min_replicates
)
}
\arguments{
\item{count_data}{a data frame}

\item{column_geneid}{a character string}

\item{sample_info}{a data frame}

\item{min_counts}{a numeric value}

\item{min_replicates}{a numeric value}
}
\value{

}
\description{
Filter count data using user thresholds for the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}
}
\keyword{internal}
