% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{.fix_ASE_diff_rslt_columns}
\alias{.fix_ASE_diff_rslt_columns}
\title{Fix column names of the SpliceWiz edgeR output.}
\usage{
.fix_ASE_diff_rslt_columns(x)
}
\arguments{
\item{x}{an output of the function \code{\link[SpliceWiz:ASE-methods]{SpliceWiz::ASE_edgeR()}}.}
}
\value{
a tibble.
}
\description{
Fix column names of the SpliceWiz edgeR output.
}
\examples{
\dontrun{
// to do
}
}
\keyword{internal}
