% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{.get_aes_heatmap_data}
\alias{.get_aes_heatmap_data}
\title{Get an input matrix for diff ASE heatmap}
\usage{
.get_aes_heatmap_data(
  se,
  event_names,
  samples,
  replicates = TRUE,
  method = "PSI",
  column_condition = "condition"
)
}
\arguments{
\item{se}{an object of class NxtSE.}

\item{event_names}{a character vector denoting valid ASE names to plot in heatmap.}

\item{samples}{a character vector denoting valid sample names to plot in the heatmap.}

\item{replicates}{logical, default TRUE, denoting whether to show sample replicates or not.}

\item{method}{a character string denoting a method for data value. Can be one of the "PSI", "logit" or "Z-Score"}

\item{column_condtion}{a name of column in se storing condition. Deafault: "condition"}
}
\value{
a data matrix
}
\description{
Get an input matrix for diff ASE heatmap
}
\examples{
se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
SpliceWiz::colData(se)$treatment <- rep(c("A", "B"), each = 3)
SpliceWiz::colData(se)$replicate <- rep(c("P","Q","R"), 2)
res_edgeR <- SpliceWiz::ASE_edgeR(se, "treatment", "A", "B", useQL = FALSE)
event_names = get_edger_diff_ase_event_names(res_edgeR,   padj_cutoff = 1)
.get_aes_heatmap_data(se, event_names , samples = c("A", "B"), column_condition ="treatment")
}
\keyword{internal}
