% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.get_all_expressed_genes}
\alias{.get_all_expressed_genes}
\title{Get list of all expressed genes from \code{x}}
\usage{
.get_all_expressed_genes(x)
}
\arguments{
\item{x}{an object of class 'parcutils'}
}
\value{
a character vector.
}
\description{
Get list of all expressed genes from \code{x}
}
\keyword{internal}
