% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.get_count_data}
\alias{.get_count_data}
\title{Prepare count data for the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}}
\usage{
.get_count_data(counts)
}
\arguments{
\item{counts}{a character sting or dataframe.}
}
\value{
a dataframe.
}
\description{
Prepare count data for the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}
}
\keyword{internal}
