% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.get_gene_groups_for_fc_scatter_plot}
\alias{.get_gene_groups_for_fc_scatter_plot}
\title{Get gene groups for fc scatter plot}
\usage{
.get_gene_groups_for_fc_scatter_plot(x, sample_comparisons, color_label)
}
\arguments{
\item{x}{an object of class parcutils}

\item{sample_comparisons}{a character vector}
}
\value{
a dataframe
}
\description{
Get gene groups for fc scatter plot
}
\keyword{internal}
