% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_realted.R
\name{.get_intron_annotations}
\alias{.get_intron_annotations}
\title{Get intron annotations}
\usage{
.get_intron_annotations(x, intron_ids)
}
\arguments{
\item{x}{an object of class irFinderSdata}

\item{intron_ids}{a character vector of intron ids for which
annotations to be retrieved.}
}
\value{
a dataframe with columns \code{chr, start, end, intron_id,
null, strand, gene_id, gene_symbol}
}
\description{
Get intron annotations
}
\examples{
example_dir <- system.file("extdata" ,"ir_data" , package="parcutils")
example_files <- fs::dir_ls(example_dir, glob = "*-IR-nondir*.txt")
names(example_files) <- stringr::str_replace(example_files , pattern = ".*/", replacement = "") \%>\%
stringr::str_replace(pattern = "-IR-nondir.txt", replacement = "")
x <- read_irfinderS_output(files = example_files,  add_prefix_chr = F)
.get_intron_annotations(x, intron_ids = paste("intron", sample(1:100, 10),sep ="_"))
}
\keyword{internal}
