% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.get_parcutils_obj_gene_names}
\alias{.get_parcutils_obj_gene_names}
\title{Get the gene names stored in the parcutils object}
\usage{
.get_parcutils_obj_gene_names(x, query_genes)
}
\arguments{
\item{x}{an object of class parcutils}

\item{query_genes}{a character vector denoting genes for which corresponding gene names to be derived from the parcutils object.}
}
\value{
a dataframe containing two columns - 1) queried genes and 2) gene names from the object parcutils.
}
\description{
Get the gene names stored in the parcutils object
}
\examples{

res = .get_parcutils_object_example()
.get_parcutils_obj_gene_names(x = res, query_genes= c("LSS", "GOLGA8S", "FXYD5", "CDC34"))

}
