% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.get_parcutils_object_example}
\alias{.get_parcutils_object_example}
\title{Obtain an example object of the class parcutils}
\usage{
.get_parcutils_object_example()
}
\value{
an object of class \code{parcutils}
}
\description{
Obtain an example object of the class parcutils
}
\keyword{internal}
