% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.get_sample_information}
\alias{.get_sample_information}
\title{Prepare sample information for the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}}
\usage{
.get_sample_information(sample_info)
}
\arguments{
\item{sample_info}{a character string or data frame.}
}
\value{
a data frame
}
\description{
Prepare sample information for the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}
}
\keyword{internal}
