% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.is_genes_present_in_obj_parcutils}
\alias{.is_genes_present_in_obj_parcutils}
\title{Check availability of genes in the object of parcutils}
\usage{
.is_genes_present_in_obj_parcutils(x, genes)
}
\arguments{
\item{x}{an object of class \code{parcutils}.}

\item{genes}{a character vector.}
}
\value{
either TRUE or abort.
}
\description{
Check availability of genes in the object of parcutils
}
\keyword{internal}
