% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.is_sample_comparsions_present_in_obj_parcutils}
\alias{.is_sample_comparsions_present_in_obj_parcutils}
\title{Check availability of sample comparisons in the object of parcutils}
\usage{
.is_sample_comparsions_present_in_obj_parcutils(x, sample_comparisons)
}
\arguments{
\item{x}{an object of class \code{parcutils}}

\item{sample_comparisons}{a character vector}
}
\value{
either TRUE or abort.
}
\description{
Check availability of sample comparisons in the object of parcutils
}
\keyword{internal}
