% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.keep_only_enriched_go}
\alias{.keep_only_enriched_go}
\title{Remove items from the list if no enriched terms found.}
\usage{
.keep_only_enriched_go(x)
}
\arguments{
\item{x}{a list containing output of GO enrichment. Each element of the list is an output of  \code{\link[clusterProfiler:enrichGO]{clusterProfiler::enrichGO()}}.}
}
\value{
a list of containing output of GO enrichment.
}
\description{
Remove items from the list if no enriched terms found.
}
\keyword{internal}
