% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.map_granges_metadata}
\alias{.map_granges_metadata}
\title{Map metadata (GC, length and seq) to GRanges object}
\usage{
.map_granges_metadata(
  x,
  bs_genome_object = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
)
}
\arguments{
\item{x}{an object of class GRanges}

\item{bs_genome_object}{an object of class BSgenome, default \code{BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38}}
}
\description{
Map metadata (GC, length and seq) to GRanges object
}
\keyword{internal}
