% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_realted.R
\name{.mark_ir_status_by_filters}
\alias{.mark_ir_status_by_filters}
\title{Filter IR results}
\usage{
.mark_ir_status_by_filters(
  x,
  min_intron_cov = 0.95,
  min_intron_depth = 5,
  minimum_splice_exact = 5,
  min_irratio = 0
)
}
\arguments{
\item{x}{an object of class irFinderSdata.}

\item{min_intron_cov}{a numeric value between 0 and 1, default 0.95, denoting minimum value for coverage cutoff}

\item{min_intron_depth}{a numeric value, default 5, denoting average sequence depth for each intron.}

\item{minimum_splice_exact}{a numeric value, default 5, denoting number of reads supporting intron splicing.}

\item{min_irratio}{a numeric value, default 0.0001, denoting ir-ratio.}
}
\value{
an object of class irFinderSdata
}
\description{
Based on filters applied, this function will checks whether each intron is retained or not.
}
\examples{
example_dir <- system.file("extdata" ,"ir_data" , package="parcutils")
example_files <- fs::dir_ls(example_dir, glob = "*-IR-nondir*.txt")
names(example_files) <- stringr::str_replace(example_files , pattern = ".*/", replacement = "") \%>\%
stringr::str_replace(pattern = "-IR-nondir.txt", replacement = "")
x <- read_irfinderS_output(files = example_files,  add_prefix_chr = TRUE)
.mark_ir_status_by_filters(x)

}
\keyword{internal}
