% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_related.R
\name{.parallel_import_bw_files}
\alias{.parallel_import_bw_files}
\title{Import several bw files using parallel processing.}
\usage{
.parallel_import_bw_files(dir_path_to_bw_files)
}
\arguments{
\item{dir_path_to_bw_files}{a character string denoting a valid path to a parent directory in which several bw files are stored in a sample wise directory. Along with a file name (without extension)  name of the corresponding sample direcory will be used as name attribute.}
}
\value{
a list of GRanges object. Each GRanges object is an output of the function \code{\link[rtracklayer:BigWigFile]{rtracklayer::import.bw()}}.
}
\description{
Import several bw files using parallel processing.
}
\details{
This function uses \code{\link[furrr:future_map]{furrr::future_map()}} to import the bw files. Following code is recommended to activate the parallel processing.

\code {future::plan(future::multisession(), workers = 50)}
\code {.parallel_import_bw_files("path/do/bw/files)}
\code {future::plan(future::sequential())}

Number of availabe threads/workers can be identified by  \code{\link[future:re-exports]{future::availableWorkers()}} and
\code{\link[future:re-exports]{future::availableCores()}}.
}
\examples{
\dontrun{
#// TODO
}
}
\keyword{{internal}}
