% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_realted.R
\name{.parcutils_assign_intron_identifier}
\alias{.parcutils_assign_intron_identifier}
\title{For each intron in the object irFinderSdata assign unique intron id}
\usage{
.parcutils_assign_intron_identifier(x)
}
\arguments{
\item{x}{an object of class irFinderSdata}
}
\value{
an object of class irFinderSdata
}
\description{
For each intron in the object irFinderSdata assign unique intron id
}
\examples{

example_dir <- system.file("extdata" ,"ir_data" , package="parcutils")
example_files <- fs::dir_ls(example_dir, glob = "*-IR-nondir*.txt")
names(example_files) <- stringr::str_replace(example_files , pattern = ".*/", replacement = "") \%>\%
stringr::str_replace(pattern = "-IR-nondir.txt", replacement = "")
x <- read_irfinderS_output(files = example_files,  add_prefix_chr = F)
.parcutils_assign_intron_identifier(x)
}
\keyword{internal}
