% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.prepare_parcutils_ir}
\alias{.prepare_parcutils_ir}
\title{Prepare object of parcutils_ir from the object of parcutils}
\usage{
.prepare_parcutils_ir(x)
}
\arguments{
\item{x}{an object of class 'parcutils'.}
}
\value{

}
\description{
Prepare object of parcutils_ir from the object of parcutils
}
\keyword{internal}
