% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.save_de_ir_results}
\alias{.save_de_ir_results}
\title{Save DE ir results in an excel file.}
\usage{
.save_de_ir_results(x, file)
}
\arguments{
\item{x}{an object of class parcutils_ir}

\item{file}{a file path.}
}
\value{
a file path.
}
\description{
Save DE ir results in an excel file.
}
\keyword{internal}
