% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.save_deg_results}
\alias{.save_deg_results}
\title{Save DEG results in an excel file.}
\usage{
.save_deg_results(x, file)
}
\arguments{
\item{x}{an object of class parcutils.}

\item{file}{a file path.}
}
\value{
output of \code{writexl::write_xlsx()}
}
\description{
Save DEG results in an excel file.
}
\keyword{internal}
