% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.simplify_go}
\alias{.simplify_go}
\title{Remove highly similar GO.}
\usage{
.simplify_go(x, similarity_cutoff = 0.9)
}
\arguments{
\item{x}{a list containing output of GO enrichment. Each element of the list is an output of  \code{\link[clusterProfiler:enrichGO]{clusterProfiler::enrichGO()}}.}

\item{similarity_cutoff}{a value between 0 and 1. Default 0.9.}
}
\value{
a list containing output of GO enrichment.
}
\description{
Remove highly similar GO.
}
\keyword{internal}
