% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_related.R
\name{.validate_a_list_of_normalised_matrix}
\alias{.validate_a_list_of_normalised_matrix}
\title{Validate a list of normalised matrix.}
\usage{
.validate_a_list_of_normalised_matrix(x)
}
\arguments{
\item{x}{a list of the object(s) of {.class normalisedMatrix}.}
}
\value{
a list of the object(s) of {.class normalisedMatrix}.
}
\description{
Validate a list of normalised matrix.
}
\examples{
\dontrun{
// TODO
}
}
