% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_realted.R
\name{.validate_irfinders_object}
\alias{.validate_irfinders_object}
\title{Check if the object belongs to class irFinderSdata.}
\usage{
.validate_irfinders_object(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Check if the object belongs to class irFinderSdata.
}
\keyword{internal}
