% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_related.R
\name{.validate_parcutils_chip}
\alias{.validate_parcutils_chip}
\title{Validate a class parcutils_chip}
\usage{
.validate_parcutils_chip(x)
}
\arguments{
\item{x}{an object of the class \code{parcutils_chip}}
}
\description{
Validate a class parcutils_chip
}
\keyword{internal}
