% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.validate_parcutils_ir_object}
\alias{.validate_parcutils_ir_object}
\title{Validate parcutils_ir object}
\usage{
.validate_parcutils_ir_object(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Validate parcutils_ir object
}
\keyword{internal}
