% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.validate_parcutils_obj}
\alias{.validate_parcutils_obj}
\title{Validate if an object is of class 'parcutils'.}
\usage{
.validate_parcutils_obj(x)
}
\arguments{
\item{x}{an object}
}
\value{
error
}
\description{
Validate if an object is of class 'parcutils'.
}
\keyword{internal}
