% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{dsr_to_tibble}
\alias{dsr_to_tibble}
\title{Convert DESeq result object in to a tibble.}
\usage{
dsr_to_tibble(x, .col_gene_id = "gene_id")
}
\arguments{
\item{x}{\code{\link[DESeq2:DESeqResults]{DESeq2::DESeqResults()}} object}
}
\value{
a data frame
}
\description{
Convert DESeq result object in to a tibble.
}
\examples{
\dontrun{
 // TO DO
}

}
\keyword{internal}
