% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{EnhancedVolcano2}
\alias{EnhancedVolcano2}
\title{Wrapper around EnhancedVolcano::EnhancedVolcano}
\usage{
EnhancedVolcano2(
  toptable,
  lab,
  x,
  y,
  pCutoff = 0.001,
  FCcutoff = 1.5,
  col_by_regul = T,
  ...
)
}
\arguments{
\item{toptable}{\code{toptable} \link{EnhancedVolcano::EnhancedVolcano}}

\item{lab}{\code{lab} \link{EnhancedVolcano::EnhancedVolcano}}

\item{x}{\code{x} \link{EnhancedVolcano::EnhancedVolcano}}

\item{y}{\code{y} \link{EnhancedVolcano::EnhancedVolcano}}

\item{pCutoff}{\code{pCutoff} \link{EnhancedVolcano::EnhancedVolcano}}

\item{FCcutoff}{\code{FCcutoff} \link{EnhancedVolcano::EnhancedVolcano}}

\item{col_by_regul}{lgl, default T, whether to color variable by gene regulation - Up, Down, Other}

\item{...}{Other parameters pass to \link{EnhancedVolcano::EnhancedVolcano}}
}
\value{
a volcano plot
}
\description{
Wrapper around EnhancedVolcano::EnhancedVolcano
}
