% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{event_region_to_coordinate}
\alias{event_region_to_coordinate}
\title{Convert event regions to genomic co-ordinates.}
\usage{
event_region_to_coordinate(event_region, prefix = "")
}
\arguments{
\item{event_region}{a character vector denoting valid event regions.}

\item{prefix}{a character string, usually "chr", to append in the chromosome names. This is useful when event names are without "chr" prefix while BSgenome contains "chr".}
}
\value{
a tibble.
}
\description{
Convert event regions to genomic co-ordinates.
}
\examples{
\dontrun{
 # // to do
}

}
