% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{event_region_to_granges}
\alias{event_region_to_granges}
\title{Convert SpliceWiz's EventRegion column to GRanges object.}
\usage{
event_region_to_granges(event_region, prefix = "")
}
\arguments{
\item{event_region}{a character vector denoting eventregion from spliceWiz out. Typically looks like this: "14:65968689-65971731/+"}

\item{prefix}{a character string denoting a prefix to be added to chromosome names. Default none. Typically the string can be "chr" or "Chr"}
}
\value{
an object of class \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges()}}
}
\description{
Convert SpliceWiz's EventRegion column to GRanges object.
}
\examples{
se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
event_region <- se@elementMetadata$Event1a \%>\% sample(10)
event_region_to_granges(event_region = event_region)

}
