% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{filter_df_by_genes}
\alias{filter_df_by_genes}
\title{Filter expression matrix (dataframe) by genes(first column).}
\usage{
filter_df_by_genes(df, genes)
}
\arguments{
\item{df}{a gene expression data frame}

\item{genes}{acharacter vector of genes to be filtred from df.}
}
\value{
a filtered dataframe.
}
\description{
Filter expression matrix (dataframe) by genes(first column).
}
\examples{
\dontrun{

# TO DO
}

}
\keyword{internal}
