% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{filter_gff}
\alias{filter_gff}
\title{Filter rows from GFF file.}
\usage{
filter_gff(
  gtf_file,
  feature_type = "gene",
  feature_biotype = "protein_coding",
  column_gene_id = "gene_id",
  column_feature_type = "type",
  column_feature_biotype = "gene_biotype"
)
}
\arguments{
\item{gtf_file}{// TO DO}

\item{feature_type}{// TO DO}

\item{feature_biotype}{// TO DO}

\item{column_gene_id}{// TO DO}

\item{column_feature_type}{// TO DO}

\item{column_feature_biotype}{// TO DO}
}
\value{
// TO DO
}
\description{
Filter rows from GFF file.
}
\examples{
\dontrun{

}

}
\keyword{internal}
