% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{fix_hm_colors}
\alias{fix_hm_colors}
\title{Given a matrix generate color scale for heatmap.}
\usage{
fix_hm_colors(hm_matrix)
}
\arguments{
\item{hm_matrix}{}
}
\value{
color for \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.
}
\description{
Given a matrix generate color scale for heatmap.
}
\keyword{internal}
