% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{get_all_named_expression_matrix}
\alias{get_all_named_expression_matrix}
\title{Get named list of gene expression matrix for all samples in x.}
\usage{
get_all_named_expression_matrix(x)
}
\arguments{
\item{x}{}
}
\value{
named list of gene expression matrix.
}
\description{
Get named list of gene expression matrix for all samples in x.
}
\keyword{internal}
