% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{get_ASE_data_matrix}
\alias{get_ASE_data_matrix}
\title{Get a data matrix (PSI, logit, z-score) for a given ASE}
\usage{
get_ASE_data_matrix(
  se,
  event_names,
  samples,
  replicates = TRUE,
  method = "PSI",
  column_condition = "condition"
)
}
\arguments{
\item{se}{an object of class NxtSE.}

\item{event_names}{a character vector denoting valid ASE names to plot in the heatmap.}

\item{samples}{a character vector denoting valid sample names to plot in the heatmap.}

\item{replicates}{logical, default TRUE, denoting whether to show sample replicates or not.}

\item{method}{a character string denoting a method for data value. Can be one of the "PSI", "logit" or "Z-Score"}

\item{column_condtion}{a name of column in se storing condition. Default: "condition"}
}
\value{
a data matrix
}
\description{
Get a data matrix (PSI, logit, z-score) for a given ASE
}
\examples{
se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
SpliceWiz::colData(se)$treatment <- rep(c("A", "B"), each = 3)
SpliceWiz::colData(se)$replicate <- rep(c("P","Q","R"), 2)
res <- run_ase_diff_analysis(x = se, test_factor = "treatment", test_nom = "A" ,test_denom = "B",  IRmode ="annotated",  cutoff_lfc = 0.6, cutoff_padj = 1, regul_based_upon = 2)
event_names = get_ASEsets_by_regulation(x = res, sample_comparisons = "A_VS_B", regul = "all") \%>\% unlist()
get_ASE_data_matrix(se, event_names , samples = c("A", "B"), column_condition ="treatment")

}
