% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_related.R
\name{get_chip_signal_heatmap}
\alias{get_chip_signal_heatmap}
\title{Generate a ChIP signal heatmap for a given sample.
#' @description This function plots the ChIP signal heatmap for the given set of target region. ChIP signals shown in the plot will be plotted as an output of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.}
\usage{
get_chip_signal_heatmap(
  path_to_bw_file,
  target_region,
  extend = 5000,
  background = NA,
  mean_mode = "w0",
  w = 50,
  smooth = TRUE,
  target_ratio = 0,
  name = NULL,
  col = NULL,
  column_title = NULL,
  axis_name = NULL,
  top_annotation = NULL,
  ...
)
}
\arguments{
\item{path_to_bw_file}{a character string denoting a path to a bw file. String can be an absolute path or the valid URL pointing to bw file.}

\item{target_region}{an object of the class \link{GenomicRanges}. Singal will be plotted against this regions.}

\item{extend}{pass to the same argument of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.}

\item{background}{pass to the same argument of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.}

\item{mean_mode}{pass to the same argument of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.}

\item{w}{pass to the same argument of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.}

\item{smooth}{pass to the same argument of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.}

\item{target_ratio}{pass to the same argument of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.}

\item{name}{pass to the same argument of \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.}

\item{col}{pass to the same argument of \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.}

\item{column_title}{pass to the same argument of \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.}

\item{axis_name}{pass to the same argument of \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.}

\item{top_annotation}{pass to the same argument of \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.}
}
\description{
Generate a ChIP signal heatmap for a given sample.
#' @description This function plots the ChIP signal heatmap for the given set of target region. ChIP signals shown in the plot will be plotted as an output of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.
}
\details{
To obtain the order of targets from the output heatmap the suggested approach is \verb{hm = ComplexHeatmap::draw(hm); ComplexHeatmap::row_order(hm)}.
}
\examples{
\dontrun{
##
}

}
