% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_related.R
\name{get_chip_signal_over_control_heatmap}
\alias{get_chip_signal_over_control_heatmap}
\title{Generate a ChIP signal heatmap of treatment over control.}
\usage{
get_chip_signal_over_control_heatmap(
  path_control_bw,
  path_treatment_bw,
  target_region = y,
  name_control = NULL,
  name_treatment = NULL,
  background = background,
  extend = extend,
  mean_mode = mean_mode,
  keep = c(0, 0.99),
  w = w,
  target_ratio = target_ratio,
  smooth = smooth,
  ...
)
}
\arguments{
\item{path_control_bw}{a character string denoting a path to a bw file. String can be an absolute path or the valid URL pointing to bw file. Singals from this file will be used as numerator to calculate the fold change.}

\item{path_treatment_bw}{a character string denoting a path to a bw file. String can be an absolute path or the valid URL pointing to bw file. Singals from this file will be used as denominator to calculate the fold change.}

\item{target_region}{an object of the class \link{GenomicRanges}. Singal will be plotted against this regions.}

\item{name_control}{a chanracter string denoting a name of the control sample. If NULL, the name will be derieved from the argument 'path_control_bw'.}

\item{name_treatment}{a chanracter string denoting a name of the treatment sample. If NULL, the name will be derieved from the argument 'path_treatment_bw'.}

\item{background}{pass to the same argument of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.}

\item{extend}{pass to the same argument of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.}

\item{mean_mode}{pass to the same argument of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.}

\item{keep}{pass to the same argument of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.}

\item{w}{pass to the same argument of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.}

\item{target_ratio}{pass to the same argument of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.}

\item{smooth}{pass to the same argument of \code{\link[EnrichedHeatmap:normalizeToMatrix]{EnrichedHeatmap::normalizeToMatrix()}}.}

\item{...}{other arguments pass to the function \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.}
}
\value{
an object of the class Heatmap from the package ComplexHeatmap.
}
\description{
This function plots the ChIP signal heatmap for the given set of target region.
ChIP signals shown in the plot are the ratio of the treatment over control. Ratio will be calculated for each bin of the target region. Prior to calculate the ratio integer 1 will be added to both numerator and denominator to avoid the Inf and NaN from the final matrix.
}
\details{
To obtain the order of targets from the output heatmap the suggested approach is \verb{hm = ComplexHeatmap::draw(hm); ComplexHeatmap::row_order(hm)}.
}
\examples{
\dontrun{

}

}
