% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{get_corr_heatbox}
\alias{get_corr_heatbox}
\title{Generate a correlation heat plot.}
\usage{
get_corr_heatbox(
  x,
  samples = NULL,
  genes = NULL,
  corr_method = "pearson",
  plot_type = "full",
  vis_method = "square",
  show_diagonal = TRUE,
  show_corr_values = FALSE,
  col_corr_values = "yellow",
  size_corr_values = 5,
  cluster_samples = FALSE,
  scale_range = NULL
)
}
\arguments{
\item{x}{an object of class parcutils.}

\item{samples}{a character vector denoting samples to plot in scatter plot, default \code{NULL}. If set to NULL all samples are accounted.}

\item{genes}{a character vector denoting genes to consider in scatter plot, default \code{NULL}. If set to NULL all genes are accounted.}

\item{corr_method}{a character string, default \code{"pearson"}, denoting a value for correlation method. Value can be one of these \code{"pearson", "kendall", "spearman"}.}

\item{plot_type}{a character string, default \code{"square"}, denoting value for plot type. Value can be one of these \code{"full", "lower", "upper"}.}

\item{vis_method}{a character string, default \code{"full"}, denoting a value type of visualization. Value can be one of these \code{"square", "circle"}.}

\item{show_diagonal}{a logical, default \code{TRUE}, denoting whether to show diagonal values or not.}

\item{show_corr_values}{a logical, default \code{TRUE}, denoting whether to show corr values or not.}

\item{col_corr_values}{a character string, default \code{"yellow"}, denoting a valid color string for corr values.}

\item{size_corr_values}{a numeric, default \code{5}, denoting a size for corr values.}

\item{cluster_samples}{a logical, default \code{TRUE}, denoting whether to cluster samples or not.}

\item{scale_range}{a numeric vector of length two denoting minimum and maximum value for the color scale, default \code{NULL}.}
}
\value{
a corr plot.
}
\description{
For all (or selected) samples this function generate a correlation heat box type plot.
}
\examples{
count_file <- system.file("extdata","toy_counts.txt" , package = "parcutils")
count_data <- readr::read_delim(count_file, delim = "\t")

sample_info <- count_data \%>\% colnames() \%>\% .[-1]  \%>\%
 tibble::tibble(samples = . , groups = rep(c("control" ,"treatment1" , "treatment2"), each = 3) )


res <- run_deseq_analysis(counts = count_data ,
                         sample_info = sample_info,
                         column_geneid = "gene_id" ,
                         group_numerator = c("treatment1", "treatment2") ,
                         group_denominator = c("control"))

get_corr_heatbox(res,samples = c("treatment1","control"),cluster_samples = FALSE,show_corr_values =T,
size_corr_values = 4)
}
