% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{get_deg}
\alias{get_deg}
\title{Perform differential expression analysis using \code{DESeq2}}
\usage{
get_deg(
  counts,
  column_geneid,
  column_samples,
  sample_info,
  group_numerator,
  group_denominator,
  delim = "\\t",
  comment_char = "#",
  min_counts = 10,
  min_replicates = 2,
  cutoff_lfc = 1,
  cutoff_pval = 0.05,
  cutoff_padj = 0.01,
  ...
)
}
\arguments{
\item{counts}{a character string providing a name of count file or a data frame of counts for each gene.
See details below to know more about format of file and count data frame.}

\item{column_geneid}{a character string denoting a column of geneid in \code{count}}

\item{column_samples}{a character vector denoting names of sample columns from \code{count}}

\item{sample_info}{a character string denoting a name of sample information file or data frame.
A file or a data frame both must have at least two columns without column names. First column denotes to samples names
and second column denotes group name for each sample in first column. For e.g.\tabular{ll}{
    \tab  \cr
   sample 1 \tab WT \cr
   sample 2 \tab WT \cr
   sample 3 \tab KO \cr
   sample 4 \tab KO \cr
}}

\item{group_numerator}{a character vector denoting sample groups to use in numerator to calculate fold change.}

\item{group_denominator}{a character vector denoting sample groups to use in denominator to calculate fold change.}

\item{delim}{a character denoting deliminator for \code{count} file. Only valid if \code{count} is a file path.}

\item{comment_char}{a character denoting comments line in count file. Only valid if \code{count} is a file path.}

\item{min_counts}{a numeric value, default 10,  denoting minimum counts for a gene to be used for differential expression analysis.}

\item{min_replicates}{a numeric value, default 2, denoting minimum samples within a group have \code{minimum_counts}.}

\item{cutoff_lfc}{a numeric value which is internally passed to \link{catagorize_diff_genes}}

\item{cutoff_pval}{a numeric value which is internally passed to \link{catagorize_diff_genes}}

\item{cutoff_padj}{a numeric value which is internally passed to \link{catagorize_diff_genes}}

\item{...}{for future use}
}
\value{
a data frame of DEG
}
\description{
This is wrapper function build upon \code{DESeq2::DESeq()} and \code{DESeq2::DESeqResults()}
to find diff genes and categories them based on various cutoffs such as p-value, padj-value, log2fc etc.
}
\examples{
\dontrun{
}
}
