% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{get_diff_ase_heatmap}
\alias{get_diff_ase_heatmap}
\title{Generate a heatmap for SpliceWiz splice events}
\usage{
get_diff_ase_heatmap(
  se,
  event_names,
  samples,
  column_condition = "condition",
  method = "PSI",
  show_replicates = TRUE,
  show_row_names = FALSE,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  show_column_dend = FALSE,
  show_row_dend = FALSE,
  ...
)
}
\arguments{
\item{se}{an object of class NxtSE.}

\item{event_names}{a character vector denoting valid spliceWiz event names}

\item{samples}{a character vector denoting valid sample names}

\item{column_condition}{a name of column in se storing condition. Deafault: "condition"}

\item{method}{one of the character strings: "Z-score", "PSI" ,"logit"}

\item{show_replicates}{logical, default TRUE, denoting whether to show sample replicates or not.}

\item{show_row_names}{logical, whether to show row names in the heatmap.}

\item{cluster_rows}{logical, whether to clusters rown in the heatmap}

\item{cluster_columns}{logical, whether to cluster column in the heatmap or not.}

\item{show_column_dend}{logical, whether to show column dendrogram in the heatmap or not.}

\item{show_row_dend}{logical, whether to show row dendrogram in the heatmap or not.}

\item{...}{other parameters to pass \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}}
}
\value{
a heatmap
}
\description{
Generate a heatmap for SpliceWiz splice events
}
\examples{
se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
SpliceWiz::colData(se)$treatment <- rep(c("A", "B"), each = 3)
SpliceWiz::colData(se)$replicate <- rep(c("P","Q","R"), 2)
res_edgeR <- SpliceWiz::ASE_edgeR(se, "treatment", "A", "B", useQL = FALSE)
event_names = get_edger_diff_ase_event_names(res_edgeR,   padj_cutoff = 1)
get_diff_ase_heatmap(se, event_names = event_names, samples = c("A" ,"B"), column_condition    = "treatment")
get_diff_ase_heatmap(se, event_names = event_names, samples = c("A" ,"B"), column_condition    = "treatment",method = "Z-score")
}
