% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{get_diff_gene_count_barplot}
\alias{get_diff_gene_count_barplot}
\title{Generate a barplot of DEG counts.}
\usage{
get_diff_gene_count_barplot(
  x,
  col_up = "#a40000",
  col_down = "#16317d",
  font_size = 12,
  show_counts = TRUE
)
}
\arguments{
\item{x}{an object of class parcutils}

\item{col_up}{a character string, default #a40000, denoting valid a color name for "Up" regulated genes.}

\item{col_down}{a character string, default #16317d, denoting valid a color name "Down" regulated genes.}

\item{font_size}{a numeric, default 12, denoting font size in the plot.}

\item{show_counts}{a logical, default \code{TRUE} denoting whether to show counts on each bar.}
}
\value{
a ggplot
}
\description{
Generate a barplot of DEG counts.
}
\examples{
res = .get_parcutils_object()
get_diff_gene_count_barplot(res, font_size = 15)
}
