% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{get_edger_diff_ase_event_names}
\alias{get_edger_diff_ase_event_names}
\title{Get diff event names from SpliceWiZ edger diff results}
\usage{
get_edger_diff_ase_event_names(
  res_edger,
  event_type = "SE",
  padj_cutoff = 0.05,
  log2fc_cutoff = 0.6
)
}
\arguments{
\item{res_edger}{an output of \code{\link[SpliceWiz:ASE-methods]{SpliceWiz::ASE_edgeR()}}}

\item{event_type}{a character string from one of the followings: SE,A5SS, A3SS, MXE, I,AFE, ALE. Default SE}

\item{padj_cutoff}{a double denoting cutoff value of padj (FDR). Default 0.05}

\item{log2fc_cutoff}{a double denoting cutoff value of log2FC (logFC). Default 0.6}
}
\value{
a character vector
}
\description{
Get diff event names from SpliceWiZ edger diff results
}
\examples{

se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
SpliceWiz::colData(se)$treatment <- rep(c("A", "B"), each = 3)
SpliceWiz::colData(se)$replicate <- rep(c("P","Q","R"), 2)
require("edgeR")
res_edgeR <- SpliceWiz::ASE_edgeR(se, "treatment", "A", "B", useQL = FALSE)
res_edgeR_QL <- SpliceWiz::ASE_edgeR(se, "treatment", "A", "B", useQL = TRUE)
get_edger_diff_ase_event_names(res_edgeR,padj_cutoff= 1)
}
