% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{get_event_annotations}
\alias{get_event_annotations}
\title{Annotate event names to event length, seq and GC.}
\usage{
get_event_annotations(se, event_names, org = "hs", prefix = "")
}
\arguments{
\item{se}{an object of class NxtSE.}

\item{event_names}{a character vector denoting valid event names.}

\item{org}{a character string "hs" or "mm".}

\item{prefix}{a character string, usually "chr", to append in the chromosome names. This is useful when event names are without "chr" prefix while BSgenome contains "chr".}
}
\value{
GenomicRanges.
}
\description{
Annotate event names to event length, seq and GC.
}
\examples{
se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
SpliceWiz::colData(se)$treatment <- rep(c("A", "B"), each = 3)
SpliceWiz::colData(se)$replicate <- rep(c("P","Q","R"), 2)
res <- run_ase_diff_analysis(x = se, test_factor = "treatment", test_nom = "A" ,test_denom = "B",  IRmode ="annotated",  cutoff_lfc = 0.6, cutoff_pval = 1, regul_based_upon = 1)
q_events <- get_ASE_by_regulation(res, sample_comparisons = "A_VS_B",regulation = "both") \%>\% names()
get_event_annotations(se = se, event_names = q_events)

}
