% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{get_fold_change_matrix}
\alias{get_fold_change_matrix}
\title{Prepare a fold change matrix}
\usage{
get_fold_change_matrix(x, sample_comparisons, genes)
}
\arguments{
\item{x}{an abject of class "parcutils". This is an output of the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}.}

\item{sample_comparisons}{a character vector denoting sample comparisons for which fold change values to be derived.}

\item{genes}{a character vector denoting gene names for which fold change values to be derived.}
}
\value{
a tbl.
}
\description{
This function returns a dataframe having first column gene names and subsequent columns are
fold change values for the comparisons passed through \code{sample_comparisons}.
}
\examples{
\dontrun{
// TO DO
}

}
