% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{get_gene_expression_box_plot}
\alias{get_gene_expression_box_plot}
\title{Generate a box plot}
\usage{
get_gene_expression_box_plot(
  x,
  samples = NULL,
  genes = NULL,
  group_replicates = F,
  convert_log2 = T
)
}
\arguments{
\item{x}{an abject of class "parcutils". This is an output of the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}.}

\item{samples}{a character vector denoting samples to plot in boxplot, default \code{NULL}. If set to NULL all samples are accounted.}

\item{genes}{a character vector denoting genes to consider for boxplot, default \code{NULL}. If set to NULL all genes are accounted.}

\item{group_replicates}{logical, default \code{FALSE}, whether to group replicates in individual plots.}

\item{convert_log2}{logical, default \code{TRUE}, whether to plot log2(noralised expression + 1) or not.}
}
\value{
an object of class ggplot2
}
\description{
Generate a box plot
}
\examples{
count_file <- system.file("extdata","toy_counts.txt" , package = "parcutils")
count_data <- readr::read_delim(count_file, delim = "\t")

sample_info <- count_data \%>\% colnames() \%>\% .[-1]  \%>\%
 tibble::tibble(samples = . , groups = rep(c("control" ,"treatment1" , "treatment2"), each = 3) )


res <- run_deseq_analysis(counts = count_data ,
                         sample_info = sample_info,
                         column_geneid = "gene_id" ,
                         group_numerator = c("treatment1", "treatment2") ,
                         group_denominator = c("control"),
                         column_samples = c("control_rep1", "treat1_rep1", "treat2_rep1", "control_rep2", "treat1_rep2", "treat2_rep2", "control_rep3", "treat1_rep3", "treat2_rep3"))

get_gene_expression_box_plot(x = res ) \%>\% print()

get_gene_expression_box_plot(x = res , group_replicates = T ) \%>\% print()

}
