% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{get_genes_from_event_name}
\alias{get_genes_from_event_name}
\title{Get gene names from spliceWiz event names}
\usage{
get_genes_from_event_name(x)
}
\arguments{
\item{x}{a character vector denoting EventNames of SpliceWiz.}
}
\value{
a character vector of gene names
}
\description{
Get gene names from spliceWiz event names
}
\examples{
se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
SpliceWiz::colData(se)$treatment <- rep(c("A", "B"), each = 3)
SpliceWiz::colData(se)$replicate <- rep(c("P","Q","R"), 2)
res_edgeR <- SpliceWiz::ASE_edgeR(se, "treatment", "A", "B", useQL = FALSE)
get_genes_from_event_name(res_edgeR$EventName)
}
