% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{get_heatmap_data}
\alias{get_heatmap_data}
\title{Get data from a heatmap in the same order.}
\usage{
get_heatmap_data(h)
}
\arguments{
\item{h}{an object of class \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}}
}
\value{
a tbl.
}
\description{
Get data from a heatmap in the same order.
}
\examples{
\dontrun{
# // TO DO
}

}
