% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_bioinfo.R
\name{get_intergenic_signals}
\alias{get_intergenic_signals}
\title{Find signal intensity of intergenic regions.}
\usage{
get_intergenic_signals(bw_file, gff_file)
}
\arguments{
\item{bw_file}{a string denoting path of .bw file}

\item{gff_file}{a string denoting path of gff file.}
}
\value{
an object of class \link{GenomicRanges} which can be exported to various file formats such as bw, bed, begraph etc.
}
\description{
given the .gff file and .bw file function finds the intergenic regions and associate signal inensity of each region
}
\examples{
\dontrun{
#' library("magrittr")
bw_file <- system.file("extdata" ,
"example.bw" ,
package = "parcutils")
gff_file <- system.file("extdata" ,
                       "C_glabrata_CBS138_version_s02-m07-r06_features.gff" ,
package = "parcutils")
oo <- parcutils::get_intergenic_signals(bw_file = bw_file,gff_file  = gff_file)
oo \%>\%
    rtracklayer::export.bed(con = file(paste("intergenic", ".bed",sep = "")))
oo \%>\%
   rtracklayer::export.bedGraph(con = file(paste("intergenic", ".bdg",sep = "")))
}
}
