% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_realted.R
\name{get_intron_master_list}
\alias{get_intron_master_list}
\title{Prepare a master list of introns}
\usage{
get_intron_master_list(
  f,
  add_meta_data = T,
  add_prefix_chr = FALSE,
  remove_prefix_chr = FALSE,
  bs_genome_object = NULL
)
}
\arguments{
\item{f}{a character string denoting irfinderS output file ending with suffix "IR-nondir".}

\item{add_meta_data}{logical, whether to map meta data or not}

\item{add_prefix_chr}{logical, whether to add prefix 'chr' in the column seqnames}

\item{remove_prefix_chr}{logical, whether to remove prefix 'chr' from the column seqnames}

\item{bs_genome_object}{an object of class BSgenome}
}
\value{
a dataframe
}
\description{
Prepare a master list of introns
}
\examples{
example_dir <- system.file("extdata" ,"ir_data" , package="parcutils")
example_files <- fs::dir_ls(example_dir, glob = "*-IR-nondir*.txt")
get_intron_master_list(f = example_files[1],add_prefix_chr = TRUE,bs_genome_object = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38)
}
