% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_realted.R
\name{get_ir_counts}
\alias{get_ir_counts}
\title{Prepare a intron reads count matrix.}
\usage{
get_ir_counts(x)
}
\arguments{
\item{x}{an object of class irFinderSdata}
}
\value{
a tbl.
}
\description{
Intron counts will be obtained from the column 'introndepth' of the IRFinder-S output.
}
\examples{

example_dir <- system.file("extdata" ,"ir_data" , package="parcutils")
example_files <- fs::dir_ls(example_dir, glob = "*-IR-nondir*.txt")
names(example_files) <- stringr::str_replace(example_files , pattern = ".*/", replacement = "") \%>\%
stringr::str_replace(pattern = "-IR-nondir.txt", replacement = "")
x <- read_irfinderS_output(files = example_files,  add_prefix_chr = F)
get_ir_counts(x)


}
