% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{get_normalised_expression_matrix}
\alias{get_normalised_expression_matrix}
\title{Prepare a matrix of normalised gene expression values}
\usage{
get_normalised_expression_matrix(
  x,
  samples = NULL,
  genes = NULL,
  summarise_replicates = FALSE,
  summarise_method = "median"
)
}
\arguments{
\item{x}{an abject of class "parcutils". This is an output of the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}.}

\item{samples}{a character vector denoting samples for which normalised gene expression values to be derived, Default NULL. If NULL it returns all samples in x}

\item{genes}{a character vector denoting gene names for which normalised gene expression values to be derived, Default NULL. If NULL it returns all samples in x.}

\item{summarise_replicates}{logical, default FALSE, indicating whether gene expression values summarised by mean or median between replicates.}

\item{summarise_method}{a character string either "mean" or "median" by which normalised gene expression values will be summarised between replicates.}
}
\value{
a tbl.
}
\description{
This function returns a dataframe having first column gene names and subsequent columns are
normalised gene expression values for the comparisons passed through sample_comparisons.
}
\examples{
\dontrun{

// TO DO
}
}
