% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{get_pairwise_corr_plot}
\alias{get_pairwise_corr_plot}
\title{Get replicates correlation scatter plot}
\usage{
get_pairwise_corr_plot(x, samples = NULL, genes = NULL)
}
\arguments{
\item{x}{an abject of class "parcutils". This is an output of the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}.}

\item{samples}{a character vector denoting samples to plot in scatter plot, default \code{NULL}. If set to NULL all samples are accounted.}

\item{genes}{a character vector denoting genes to consider in scatter plot, default \code{NULL}. If set to NULL all genes are accounted.#'}
}
\value{
a named list.
}
\description{
Get replicates correlation scatter plot
}
\examples{
\dontrun{
// TO DO
}


}
