% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{get_pca_plot}
\alias{get_pca_plot}
\title{Generate a PCA plot.}
\usage{
get_pca_plot(
  x,
  samples = NULL,
  genes = NULL,
  circle_size = 10,
  label_replicates = FALSE
)
}
\arguments{
\item{x}{an abject of class "parcutils". This is an output of the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}.}

\item{samples}{a character vector denoting samples to plot in PCA plot, default \code{NULL}. If set to NULL all samples are accounted.}

\item{genes}{a character vector denoting genes to consider for PCA plot, default \code{NULL}. If set to NULL all genes are accounted.}

\item{circle_size}{a numeric value,  default  10, denoting size of the circles in PCA plot.}

\item{show_replicates}{a logical, degfault FALSE, denoting whether to label replicates in the PCA plot.}
}
\value{
an object of ggplot2.
}
\description{
Generate a PCA plot.
}
\examples{
\dontrun{
 // TO DO
}


}
