% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_bioinfo.R
\name{get_rpkm}
\alias{get_rpkm}
\title{Calculate RPKM (Reads per kilobase per million).}
\usage{
get_rpkm(counts, lengths)
}
\arguments{
\item{counts}{a numeric vector giving counts of reads aligned against genomic intervals.}

\item{lengths}{a numeric vector giving length of genomic intervals.}
}
\value{
a numeric vector giving RPKM
}
\description{
given the raw counts and length of genomic features this function returns normalized counts as reads per kilo-base per million  (RPKM).
}
\keyword{internal}
