% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{get_splicewiz_event_counts}
\alias{get_splicewiz_event_counts}
\title{Generate a barplot showing counts of alternate splice events from the object NxtSe.}
\usage{
get_splicewiz_event_counts(x, elem_text_size = 15, text_count_size = 10)
}
\arguments{
\item{x}{an object of class \code{NxtSE}. Usually \code{se} or \code{se.filtered}.}

\item{elem_text_size}{an integer denoting size of the various text elements in the plot.}

\item{text_count_size}{an integer denoting size of the counts mentioned above each bar.}
}
\value{
ggplot.
}
\description{
Generate a barplot showing counts of alternate splice events from the object NxtSe.
}
\examples{
se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
get_splicewiz_event_counts(se, elem_text_size = 15, text_count_size = 4)


}
\keyword{internal}
