% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_bioinfo.R
\name{get_star_align_log_summary}
\alias{get_star_align_log_summary}
\title{Get alignment summary from the output of STAR program.}
\usage{
get_star_align_log_summary(log_file)
}
\arguments{
\item{log_file}{a string denoting a path to *Log.final.out file generated by STAR aligner.}
}
\value{
a dataframe.
}
\description{
STAR is widely used alignment software program for RNA-seq and other high throughput genomics data.
Once the alignment finishes obvious question is to know what percentage of the reads mapped to the reference.
These information is logged in *Log.final.out file which is one of the several files STAR generates at the end of alignment.
This function parse *Log.final.out file and extract the mapping statistics in a dataframe format.
}
\examples{

star_align_log_file <- system.file("extdata" , "toy_star_align_Log.final.out" , package = "parcutils")
x =  get_star_align_log_summary(log_file = star_align_log_file)
print(x)


}
