% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_related.R
\name{get_thee_prime_flank_motif}
\alias{get_thee_prime_flank_motif}
\title{Get a seqlogo for flanking region across three prime end of the genomic range.}
\usage{
get_thee_prime_flank_motif(
  x,
  y = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
  extend = 30
)
}
\arguments{
\item{x}{an object of the class GenomicRanges}

\item{y}{an object of class BSgenome, default \code{BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38}}

\item{extend}{an integer, default 30, denoting number of basepairs to extend from three prime end of the range.}
}
\value{
a list containing ggseqlogo and sequences used to generate the logo.
}
\description{
given an object of  \code{GenomicRanges} the function generates a seqlogo
of the region flanked around three prime end.
}
\examples{
\dontrun{

}

}
