% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_bioinfo.R
\name{get_tpm}
\alias{get_tpm}
\title{Calculate TPM (Tags Per Million).}
\usage{
get_tpm(counts, lengths)
}
\arguments{
\item{counts}{a numeric vector giving counts of reads aligned against genomic intervals.}

\item{lengths}{a numeric vector giving length of genomic intervals.}
}
\value{
a numeric vector giving TPM
}
\description{
given the raw counts and length of genomic features this function returns normalized counts as tags per million (TPM).
}
\keyword{internal}
