% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{get_volcano_plot}
\alias{get_volcano_plot}
\title{Generate a volcano plot.}
\usage{
get_volcano_plot(
  x,
  sample_comparison,
  log2fc_cutoff = 1,
  pval_cutoff = 0.05,
  genes_to_display = NULL,
  lab_size = 3,
  point_size = 1,
  col_up = "#a40000",
  col_down = "#007e2f",
  col_other = "grey",
  ...
)
}
\arguments{
\item{x}{an abject of class "parcutils". This is an output of the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}.}

\item{sample_comparison}{a character string denoting a valid differatnial gene comparison. Possible comparisons can be found from x$comp.}

\item{log2fc_cutoff}{a numeric value, default 1.}

\item{pval_cutoff}{a numeric value, default 0.05.}

\item{genes_to_display}{a character vector of the genes to display in volcano plot, default NULL, displays non overlapping genes.}

\item{lab_size}{a numeric value, default 3, denoting size of the lables.}

\item{point_size}{a numeric value, default 1, denoting size of the points}

\item{col_up}{a character string, default "a40000", denoting valid color code for up regulated genes.}

\item{col_down}{a character string, default "007e2f", denoting valid color code for down regulated genes.}

\item{col_other}{a character string, default "grey", denoting valid color code for other than up and down regulated genes.}

\item{...}{other parameters to be passed to \code{\link[EnhancedVolcano:EnhancedVolcano]{EnhancedVolcano::EnhancedVolcano()}}}
}
\value{
ggplot
}
\description{
Generate a volcano plot.
}
\examples{
\dontrun{

// TO DO
}

}
