% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{ggsave2}
\alias{ggsave2}
\title{Wrapper around ggsave.}
\usage{
ggsave2(plot, path, file_name, height = 5, width = 5, ...)
}
\arguments{
\item{plot}{a ggplot}

\item{path}{directory path excluding file name}

\item{file_name}{a file name with extension}

\item{height}{device height}

\item{width}{device width}

\item{...}{other arguments pass to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\value{
save a plot in a directory.
}
\description{
save a ggplot in a directory
}
