% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{group_replicates_by_sample}
\alias{group_replicates_by_sample}
\title{Group replicates by samples.}
\usage{
group_replicates_by_sample(x)
}
\arguments{
\item{x}{an abject of class "parcutils". This is an output of the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}.}
}
\value{
a tbl.
}
\description{
Group replicates by samples.
}
\examples{
\dontrun{
// TO DO
}
}
\keyword{internal.}
