% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_related.R
\name{import_topn_bed_features}
\alias{import_topn_bed_features}
\title{Import top features (ranked by score) from a bed file.}
\usage{
import_topn_bed_features(bed_feature_file, topn = 5000, center = TRUE, ...)
}
\arguments{
\item{bed_feature_file}{a character string denoting a valid bed file.}

\item{topn}{a numeric value, default 5000, denoting number of top features to keep. Features are ranked by a column  {.code score}.}

\item{center}{logical, denoting whether to align each feature by central position. If TRUE, default, it will return a ranges of single nucleotide (width = 0) denting a central position of each feature.}

\item{...}{Other arguments pass to the function \code{\link[rtracklayer:BEDFile-class]{rtracklayer::import.bed()}}.}
}
\value{
an object of the {.class GRanges} containing {.arg topn} features.
}
\description{
Import top features (ranked by score) from a bed file.
}
\examples{
\dontrun{

// TO DO.
}
}
