% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_related.R
\name{make_parcutils_chip}
\alias{make_parcutils_chip}
\title{Construct a parcutils_chip object}
\usage{
make_parcutils_chip(x, row_ranges)
}
\arguments{
\item{x}{a list of the object(s) of {.cls normalisedMatrix}.}

\item{row_ranges}{an object of the class GenomicRanges. Each range must be identified by an unique id stored in a column \code{name},}
}
\value{
an object of the class parcutils_chip. Class parcutils_chip extends to RangedSummarizedExperiment. It contains data in the below format
\itemize{
\item{"assay"} {These is the object of NormalisedMatrix for the sample.}
\item{"columns"} {Columns are the columns in each assays. In this case they are the bins described in NormalisedMatrix.}
\item{"rowData"} {RPM value for each feature (row) across samples. }
}
}
\description{
This function creates an object of the  {.cls parcutils_chip.} parcutils_chip is an extension of the the {.cls RangedSummarizedExperiment}.
This object make sure that it contains necessary elements to perform downstream chipseq analysis. Downstream exploratory and visualization functions will be based on this object.
}
\details{
RPM is calculated by the function RowSums applied on the each row of the
{.arg normalized_matrix}.
}
\examples{

signal = GRanges(seqnames = "chr1",
ranges = IRanges(start = c(1, 4, 7, 11, 14, 17, 21, 24, 27),
                 end = c(2, 5, 8, 12, 15, 18, 22, 25, 28)),
score = c(1, 2, 3, 1, 2, 3, 1, 2, 3))
target = GRanges(seqnames = "chr1", ranges = IRanges(start = 10, end = 20), name = "feat_1")
x = normalizeToMatrix(signal, target, extend = 10, w = 2)
 rownames(x) <- "feat_1"
make_parcutils_chip(x, target)

}
