% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_realted.R
\name{map_intron_meta_data}
\alias{map_intron_meta_data}
\title{Map metadata to irfinder-s output}
\usage{
map_intron_meta_data(x, bs_genome_object = BSgenome.Hsapiens.UCSC.hg38)
}
\arguments{
\item{x}{an object of class irFinderSdata}

\item{bs_genome_object}{an object of class BSgenome}
}
\value{
a dataframe
}
\description{
This function allows mapping DNA sequence, GC content and intron length to irfinder-s output
}
\examples{
example_dir <- system.file("extdata" ,"ir_data" , package="parcutils")
example_files <- fs::dir_ls(example_dir, glob = "*-IR-nondir*.txt")
names(example_files) <- stringr::str_replace(example_files , pattern = ".*/", replacement = "") \%>\%
stringr::str_replace(pattern = "-IR-nondir.txt", replacement = "")
x <- read_irfinderS_output(files = example_files[1],  add_prefix_chr = TRUE)
map_intron_meta_data(x = x,  bs_genome_object = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38)
}
