% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_bioinfo.R
\name{named_group_split}
\alias{named_group_split}
\title{split grouped data with names. A wrapper around dplyr::group_by()
ref: https://github.com/tidyverse/dplyr/issues/4223}
\usage{
named_group_split(.tbl, ..., keep_order = T)
}
\arguments{
\item{.tbl}{a data frame}

\item{...}{arguments pass to dplyr::group_by()}

\item{keep_order}{logical, default TRUE, whether to maintain original order of the groups.}
}
\value{
a named list
}
\description{
split grouped data with names. A wrapper around dplyr::group_by()
ref: https://github.com/tidyverse/dplyr/issues/4223
}
\examples{
\dontrun{
a <- tibble::tibble(x = 1:5, y = sample(letters[1:5]))
a  \%>\% named_group_split(y)
a  \%>\% named_group_split(y , keep_order = F)
}


}
