% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_bioinfo.R
\name{normalise_counts}
\alias{normalise_counts}
\title{Derive normalized counts from raw read numbers}
\usage{
normalise_counts(x, .vars, method = "TPM")
}
\arguments{
\item{x}{A dataframe of raw counts along with mandatory columns which are \code{GeneID, Chr, Start, End, Strand, Length}.}

\item{.vars}{A character vector containing columns from \code{x}. Normalization will be performed on these columns.}

\item{method}{A character string, default TPM. Choices are one of TPM, RPKM.}
}
\value{
A dataframe with all mandatory columns along with columns mentioned in \code{.vars}. Remaining columns from \code{x} will be dropped.
}
\description{
This function allows converting raw read counts into normalized counts either by method Tags Per Million (TPM) or Reads Per Kilo-base Per Million (RPKM)
}
\details{
implemntaion of RPKM and TPM can be seen in functions \link[parcutils]{get_tpm} and \link[parcutils]{get_rpkm} respectively.
}
\examples{
\dontrun{
set.seed(123)
tt <- tibble::tibble(GeneID = c(paste("Gene_",1:5,sep = "")),
                     Chr = "Chr1",
                     Start = sample(1:100, 5),
                     End = sample(100:200,5),
                     Strand = sample(c("+" ,"-"), 5, replace = T),
                     Length = (End - Start )+ 1 )

tt \%<>\% dplyr::mutate(sample_1 = sample(c(1:100),5)*10 ,
                      sample_2 = sample(c(1:100),5)*10 ,
                      sample_3 = sample(c(1:100),5)*100,
                      sample_4 = sample(c(1:100),5)*100 )


normalise_counts(x = tt, .vars = c("sample_1","sample_2") ,method = "TPM")
normalise_counts(x = tt, .vars = c("sample_1","sample_2") ,method = "RPKM")
}
}
