% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_related.R
\docType{class}
\name{parcutils_chip-class}
\alias{parcutils_chip-class}
\alias{parcutils_chip}
\title{Set class parcutils_chip}
\value{
an object of class parcutils_chip
}
\description{
Set class parcutils_chip
}
\examples{
\dontrun{
# // todo
}
}
\keyword{internal}
