% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_related.R
\name{plot_chip_pairwise_correlation_heatmap}
\alias{plot_chip_pairwise_correlation_heatmap}
\title{Generate a pairwise correlation for chipseq data.}
\usage{
plot_chip_pairwise_correlation_heatmap(
  x,
  samples = NULL,
  peaks = NULL,
  rename_samples = NULL,
  scale_min = 0,
  scale_max = 1,
  col = c("blue", "white", "red"),
  cor_method = c("pearson"),
  plot_title = "title",
  ...
)
}
\arguments{
\item{x}{an object of the class \code{parcutils_chip}.}

\item{samples}{a character vector denoting samples to show in the plot. Values must be a subset of \code{names(assays(x))}.}

\item{peaks}{a character vector denoting peaks (observations) to use for correlation. Values must be a subset of \code{rownames(RowData(x))}.}

\item{rename_samples}{a character vector denoting sample names to show in the plot. Values must be of same order and length of values given as an argument \code{sample}.}

\item{scale_min}{a numeric value denoting a minimum value for the scale.}

\item{scale_max}{a numeric value denoting a maximum value for the scale.}

\item{col}{a character string of valid colors. Number of colors are not restricted.}

\item{cor_method}{a character string denoting a method to calculate correlation values. Values can be one of \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}.}

\item{plot_title}{a character string denoting a plot title.}

\item{...}{other arguments pass to the function \code{\link[ggcorrplot:ggcorrplot]{ggcorrplot::ggcorrplot()}}.}
}
\value{
a plot
}
\description{
Given an object of the class \code{parcutils_chip} it generates correlation heatmap for selected samples and observations (peaks).
}
\details{
correlation is calculated based on RPM value of each feature. RPM is simply sum of values across the feature i.e. rowsum.
}
\examples{
\dontrun{

}

}
