% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{plot_deASE_venn}
\alias{plot_deASE_venn}
\title{Plot a venn diagram for diff ASE (using event names or gene names) between multiple comparison.}
\usage{
plot_deASE_venn(
  x,
  sample_comparisons,
  regulation = "both",
  fill_color = NULL,
  event_type = NULL,
  ovelap_genes = TRUE,
  ...
)
}
\arguments{
\item{x}{an abject of class "parcutils_ase". This is an output of the function \code{\link[=run_ase_diff_analysis]{run_ase_diff_analysis()}}.}

\item{sample_comparisons}{a character vector denoting  sample comparisons for which ASE to be obtained.}

\item{regulation}{a character string, default \code{both}. Values can be one of the \code{up}, \code{down}, \code{both}, \code{other}, \code{all}.}

\item{fill_color}{an argument pass to the function \link[ggvenn:ggvenn]{ggvenn::ggvenn}.}

\item{event_type}{a character string denoting event type for which one of the \code{up}, \code{down}, \code{both}, \code{other} or \code{all} events to be returned based.
Choices are: \code{IR}, \code{SE}, \code{AFE}, \code{ALE}, \code{A3SS}, \code{A5SS} and \code{MXE.} Default NULL, uses all event types.}

\item{ovelap_genes}{logical, default \code{TRUE}, instead of event_names, overlap gene names.}

\item{...}{other arguments pass to the function \link[ggvenn:ggvenn]{ggvenn::ggvenn}}
}
\value{
a venn diagram
}
\description{
Plot a venn diagram for diff ASE (using event names or gene names) between multiple comparison.
}
\examples{

se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
SpliceWiz::colData(se)$treatment <- rep(c("A", "B"), each = 3)
SpliceWiz::colData(se)$replicate <- rep(c("P","Q","R"), 2)
res <- run_ase_diff_analysis(x = se, test_factor = "replicate", test_nom = c("P","Q","R") ,test_denom = c("Q","R","P"),  IRmode ="annotated",  cutoff_lfc = 0.6, cutoff_pval = 1, regul_based_upon = 1)
plot_deASE_venn(x = res, sample_comparisons = c("P_VS_Q","Q_VS_R","R_VS_P"),  event_type = "SE", regulation = "both", ovelap_genes = TRUE)
plot_deASE_venn(x = res, sample_comparisons = c("P_VS_Q","Q_VS_R","R_VS_P"),  event_type = "SE", regulation = "up", ovelap_genes = TRUE)
plot_deASE_venn(x = res, sample_comparisons = c("P_VS_Q","Q_VS_R","R_VS_P"),  event_type = "A3SS", regulation = "both", ovelap_genes = TRUE)

}
