% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{plot_deg_venn}
\alias{plot_deg_venn}
\title{Generate a venn diagram showing overlap between DE genes.}
\usage{
plot_deg_venn(x, sample_comparisons, regulation = "up", fill_color = NULL, ...)
}
\arguments{
\item{x}{an abject of class \code{parcutils}. This is an output of the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}.}

\item{sample_comparisons}{a character vector of length 2 denoting DE comparisons for which venn diagram to plot.}

\item{regulation}{a character string from \code{up, down, both}. If \code{up} or \code{down} respective gene sets from the \code{sample_comparisons} will be used for venn diagram. If \code{both}, both \code{up} and \code{down}
will be from each \code{sample_comparisons} will be used for venn diagram.}

\item{fill_color}{an argument pass to the function \link[ggvenn:ggvenn]{ggvenn::ggvenn}.}

\item{...}{other arguments pass to the function \link[ggvenn:ggvenn]{ggvenn::ggvenn}}
}
\value{
ggplot
}
\description{
Generate a venn diagram showing overlap between DE genes.
}
\examples{
x <- parcutils:::.get_parcutils_object_example()
sample_comparisons <- c("treatment1_VS_control","treatment2_VS_control")
plot_deg_venn(x = x, sample_comparisons = sample_comparisons , regulation = "up")
plot_deg_venn(x = x, sample_comparisons = sample_comparisons , regulation = "down")
plot_deg_venn(x = x, sample_comparisons = sample_comparisons , regulation = "both")
}
