% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{plot_regions_relative_to_reference}
\alias{plot_regions_relative_to_reference}
\title{Generate segment plot to visualize genomic regions related to reference regions.}
\usage{
plot_regions_relative_to_reference(query, reference)
}
\arguments{
\item{query}{an object of \code{\link[GenomicRanges:genomic-range-squeezers]{GenomicRanges::granges()}} to be visualized relative to reference}

\item{reference}{an object of \code{\link[GenomicRanges:genomic-range-squeezers]{GenomicRanges::granges()}} against which \code{query} regions should be mapped.}
}
\value{
a list containing segment plot and plot data
}
\description{
Generate segment plot to visualize genomic regions related to reference regions.
}
\examples{
\dontrun{
ref_file <- system.file("extdata", "Homo_sapiens.GRCh38.101.gtf.gz" ,package = "parcutils")
ref_data <- rtracklayer::import(ref_file)  \%>\%
 tibble::as_tibble() \%>\%
 dplyr::mutate(seqnames = stringr::str_c("chr",seqnames ,sep = "")) \%>\% plyranges::as_granges()
 ref_data_gr <-  ref_data \%>\%
dplyr::filter(type == "gene" & gene_biotype == "protein_coding")
query_regions_gr <- ref_data \%>\% dplyr::filter(type == "Selenocysteine")  \%>\% sample(5000)

oo <- plot_regions_relative_to_reference(query = query_regions_gr, reference  =  ref_data_gr)
 oo$plot
oo$data \%>\% ggplot2::ggplot(aes(x = q_relt_start)) + geom_density()
}
}
