% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.parcutils}
\alias{print.parcutils}
\title{S3 print for object \code{parcutils}}
\usage{
\method{print}{parcutils}(object)
}
\arguments{
\item{object}{}
}
\value{
print default
}
\description{
S3 print for object \code{parcutils}
}
\keyword{internal}
