% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.parcutils_ase}
\alias{print.parcutils_ase}
\title{S3 print for object \code{parcutils_ase}}
\usage{
\method{print}{parcutils_ase}(object)
}
\arguments{
\item{object}{}
}
\description{
S3 print for object \code{parcutils_ase}
}
\keyword{internal}
