% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.parcutils_ir}
\alias{print.parcutils_ir}
\title{S3 print for object \code{parcutils_ir}}
\usage{
\method{print}{parcutils_ir}(object)
}
\arguments{
\item{object}{}
}
\description{
S3 print for object \code{parcutils_ir}
}
\keyword{internal}
