% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print}
\alias{print}
\title{S3 print for object \code{parcutils}}
\usage{
print(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
S3 print for object \code{parcutils}
}
\keyword{internal}
