% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_realted.R
\name{read_irfinderS_output}
\alias{read_irfinderS_output}
\title{Read IR-finderS output}
\usage{
read_irfinderS_output(files, add_prefix_chr = TRUE, remove_prefix_chr = FALSE)
}
\arguments{
\item{files}{a character vector denoting irfinderS output file(s) ending with suffix "IR-nondir".}

\item{add_prefix_chr}{logical, whether to add prefix 'chr' in the column seq names}

\item{remove_prefix_chr}{logical, whether to remove prefix 'chr' from the column seq names
\itemize{
\item chr
\item start
\item end
\item name
\item null
\item strand
\item coverage
\item introndepth
\item spliceleft
\item spliceright
\item spliceexact
\item irratio
\item warnings
}}
}
\description{
Read IR-finderS output
}
\examples{
example_dir <- system.file("extdata" ,"ir_data" , package="parcutils")
example_files <- fs::dir_ls(example_dir, glob = "*-IR-nondir*.txt")
names(example_files) <- stringr::str_replace(example_files , pattern = ".*/", replacement = "") \%>\%
stringr::str_replace(pattern = "-IR-nondir.txt", replacement = "")
read_irfinderS_output(files = example_files,  add_prefix_chr = FALSE)
}
