% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{run_ase_diff_analysis}
\alias{run_ase_diff_analysis}
\title{Perform ase diff analysis using edgeR for multiple comparison.}
\usage{
run_ase_diff_analysis(
  x,
  test_nom,
  test_denom,
  test_factor = "condition",
  cutoff_lfc = 1,
  cutoff_pval = 0.05,
  cutoff_padj = 0.01,
  regul_based_upon = 1,
  ...
)
}
\arguments{
\item{x}{an object of class \code{NxtSe}.}

\item{test_nom, test_denom}{A pair of vectors, usually the same length, denoting nominator and denominator conditions to test for differential ASE, respectively.
Usually the "treatment" condition is used in nominator and "control" is used in the denominator.}

\item{test_factor}{refer to the argument \code{test_factor} in \code{\link[SpliceWiz:ASE-methods]{SpliceWiz::ASE_edgeR()}}.}

\item{cutoff_lfc}{minimal threshold for log2fold change, default 1 (2 fold).}

\item{cutoff_pval}{minimal threshold for pvalue, default 0.05. P-value threshold will be applied only when
\code{regul_based_upon} is either 1 or 3.}

\item{cutoff_padj}{minimal threshold for Padj, default 0.01. Padj threshold will be applied only when
\code{regul_based_upon} is either 2 or 3.}

\item{regul_based_upon}{one of the numeric choices  1, 2, or 3. Default 1 i.e. categorized diff. ASE by pvalue and log2fc.
\subsection{if 1 ...}{
\itemize{
\item Up : log2fc >= cutoff_lfc & pvalue <= cutoff_pval
\item Down : log2fc  <= (-1) * cutoff_lfc & pvalue <= cutoff_pval
\item Other : remaining genes
}
}

\subsection{if 2 ...}{
\itemize{
\item Up : log2fc >= cutoff_lfc & padj <= cutoff_padj
\item Down : log2fc  <= (-1) * cutoff_lfc & padj <= cutoff_padj
\item Other : remaining genes
}
}

\subsection{if 3 ...}{
\itemize{
\item Up : log2fc >= cutoff_lfc & pvalue <= cutoff_pval & padj <= cutoff_padj
\item Down : log2fc  <= (-1) * cutoff_lfc & pvalue <= cutoff_pval & padj <= cutoff_padj
\item Other : remaining genes
}
}}

\item{...}{Other parameters passed to \code{\link[SpliceWiz:ASE-methods]{SpliceWiz::ASE_edgeR()}}.}
}
\value{
an object of class parcutils_se.
}
\description{
Perform ase diff analysis using edgeR for multiple comparison.
}
\examples{
se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
SpliceWiz::colData(se)$treatment <- rep(c("A", "B"), each = 3)
SpliceWiz::colData(se)$replicate <- rep(c("P","Q","R"), 2)
run_ase_diff_analysis(x = se, test_factor = "treatment", test_nom = "A" ,test_denom = "B",  IRmode ="annotated")
run_ase_diff_analysis(x = se, test_factor = "treatment", test_nom = "A" ,test_denom = "B",  IRmode ="annotated",  cutoff_lfc = 0.6, cutoff_padj = 1, regul_based_upon = 2)

}
