% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_realted.R
\name{save_de_results}
\alias{save_de_results}
\title{Save differential expression results in an excel file.}
\usage{
save_de_results(x, file)
}
\arguments{
\item{x}{an object of class parcutils.}

\item{file}{a character string denoting a file name WITHOUT file extension. (e.g. "out_file" or "path/to/save/data/out_file")}
}
\value{
an absolute file path denoting a file in which data saved.
}
\description{
All DE comparison from the \code{x} will be saved in an excel file.
Each tab within the excel file denotes an individual DE comparison.
}
\examples{
example_dir <- system.file("extdata" ,"ir_data" , package="parcutils")
example_files <- fs::dir_ls(example_dir, glob = "*-IR-nondir*.txt")
names(example_files) <- stringr::str_replace(example_files , pattern = ".*/", replacement = "") \%>\%
stringr::str_replace(pattern = "-IR-nondir.txt", replacement = "")
x <- read_irfinderS_output(files = example_files,  add_prefix_chr = FALSE)
ir_sample_info <- names(example_files) \%>\%
tibble::tibble(samples = . , groups = rep(c("control", "treatment"), each = 2) )
out <- run_deseq_analysis_ir(x = x, sample_info = ir_sample_info,
min_intron_depth = 2, regul_based_upon =1, cutoff_pval = 0.01)
save_de_results(out, file = fs::file_temp())
}
