% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{save_go_plots}
\alias{save_go_plots}
\title{Save GO plot(s) in .pdf files.}
\usage{
save_go_plots(
  x,
  output_dir = ".",
  output_file_suffix = "go_plot",
  height = 10,
  width = 10
)
}
\arguments{
\item{x}{x an object of class 'parcutils_go_results' which can be generated using the function \code{\link[=get_go_emap_plot]{get_go_emap_plot()}}}

\item{output_dir}{a character string, default "./", denoting a path to output file.}

\item{output_file_suffix}{a character string, default "go_plot", denoting a suffix for go plot pdf files.
If default used, the name of output file will be <name_of_DE_comparison>_<up/down>_go_plot.pdf.
For e.g.  for \code{up} regulated genes from the comparison treatment_VS_control the file name will be}

\item{height}{plot height in inch, default 10.}

\item{width}{plot width in inch, default 10.
\code{treatment_VS_control_up_go_plot.pdf}}
}
\value{
It saves file into a dir.
}
\description{
Save GO plot(s) in .pdf files.
}
