% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{save_go_results}
\alias{save_go_results}
\title{Save GO results in a excel file.}
\usage{
save_go_results(x, output_dir = ".", output_file_name = "go_outout")
}
\arguments{
\item{x}{an object of class 'parcutils_go_results' which can be generated using the function \code{\link[=get_go_emap_plot]{get_go_emap_plot()}}}

\item{output_dir}{a character string, default "./", denoting a path to output file.}

\item{output_file_name}{a character string, default "go_output", denoting an output file name.}
}
\value{
a list with two elements - 1) go results and 2) go emap plots.
}
\description{
For each geneset, enrich GO terms and related data will be saved in an individual tab within a single excel file.
}
