% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_realted.R
\name{select_cols_irfinderS_output}
\alias{select_cols_irfinderS_output}
\title{Subset columns from IRfinder-S output}
\usage{
select_cols_irfinderS_output(
  x,
  keep_columns = c("chr", "start", "end", "name", "null", "strand", "coverage",
    "introndepth", "spliceleft", "spliceright", "spliceexact", "irratio", "warnings")
)
}
\arguments{
\item{x}{an object of class irFinderSdata}

\item{keep_columns}{a character vector denoting column names to keep in the output dataframe
\itemize{
\item chr
\item start
\item end
\item name
\item null
\item strand
\item coverage
\item introndepth
\item spliceleft
\item spliceright
\item spliceexact
\item irratio
\item warnings
}}
}
\value{

}
\description{
Subset columns from IRfinder-S output
}
\examples{
example_dir <- system.file("extdata" ,"ir_data" , package="parcutils")
example_files <- fs::dir_ls(example_dir, glob = "*-IR-nondir*.txt")
names(example_files) <- stringr::str_replace(example_files , pattern = ".*/", replacement = "") \%>\%
stringr::str_replace(pattern = "-IR-nondir.txt", replacement = "")
x <- read_irfinderS_output(files = example_files,  add_prefix_chr = FALSE)
select_cols_irfinderS_output(x)
}
