% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{spliceWiz_plot_event_counts}
\alias{spliceWiz_plot_event_counts}
\title{Generate a barplot showing counts of splice events}
\usage{
spliceWiz_plot_event_counts(x)
}
\arguments{
\item{x}{an object of class \code{NxtSE}}
}
\value{
ggplot.
}
\description{
Generate a barplot showing counts of splice events
}
\examples{
\dontrun{

}
}
