% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{validata_parcutils_obj}
\alias{validata_parcutils_obj}
\title{Validate if an object is of class 'parcutils'.}
\usage{
validata_parcutils_obj(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Validate if an object is of class 'parcutils'.
}
\keyword{internal}
