% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_realted.R
\name{.map_granges_metadata}
\alias{.map_granges_metadata}
\title{map metadata (GC, length and seq) to the GRanges object}
\usage{
.map_granges_metadata(x, bs_genome_object = BSgenome.Hsapiens.UCSC.hg38)
}
\arguments{
\item{x}{an object of class GRanges}

\item{bs_genome_object}{an object of class BSgenome}
}
\value{

}
\description{
map metadata (GC, length and seq) to the GRanges object
}
\keyword{internal}
