% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.validate_irfinders_object}
\alias{.validate_irfinders_object}
\title{Check if the object belongs to class irFinderSdata.}
\usage{
.validate_irfinders_object(x)
}
\arguments{
\item{x}{an object to be tested for class 'irFinderSdata'}
}
\value{
error.
}
\description{
Check if the object belongs to class irFinderSdata.
}
\keyword{internal}
