% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{EnhancedVolcano2}
\alias{EnhancedVolcano2}
\title{Wrapper around EnhancedVolcano::EnhancedVolcano}
\usage{
EnhancedVolcano2(
  toptable,
  lab,
  x,
  y,
  pCutoff = 0.001,
  FCcutoff = 1.5,
  col_by_regul = T,
  col_up = "#b2182b",
  col_down = "#2166ac",
  col_others = "#e0e0e0",
  ...
)
}
\arguments{
\item{toptable}{\code{toptable} \code{\link[EnhancedVolcano:EnhancedVolcano]{EnhancedVolcano::EnhancedVolcano()}}}

\item{lab}{\code{lab} \code{\link[EnhancedVolcano:EnhancedVolcano]{EnhancedVolcano::EnhancedVolcano()}}}

\item{x}{\code{x} \code{\link[EnhancedVolcano:EnhancedVolcano]{EnhancedVolcano::EnhancedVolcano()}}}

\item{y}{\code{y} \code{\link[EnhancedVolcano:EnhancedVolcano]{EnhancedVolcano::EnhancedVolcano()}}}

\item{pCutoff}{\code{pCutoff} \code{\link[EnhancedVolcano:EnhancedVolcano]{EnhancedVolcano::EnhancedVolcano()}}}

\item{FCcutoff}{\code{FCcutoff} \code{\link[EnhancedVolcano:EnhancedVolcano]{EnhancedVolcano::EnhancedVolcano()}}}

\item{col_by_regul}{lgl, default T, whether to color variable by gene regulation - Up, Down, Other}

\item{col_up}{a character string denoting color for up genes, works only when col_by_regul = T}

\item{col_down}{a character string denoting color for down genes, works only when col_by_regul = T.}

\item{col_others}{a character string denoting color for other genes, works only when col_by_regul = T.}

\item{...}{Other parameters pass to \code{\link[EnhancedVolcano:EnhancedVolcano]{EnhancedVolcano::EnhancedVolcano()}}}
}
\value{
a volcano plot
}
\description{
Wrapper around EnhancedVolcano::EnhancedVolcano
}
