% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{enrichMsigDB}
\alias{enrichMsigDB}
\title{Perform over representation analysis (ORA) against Molecular Signature Database (MsigDB) .}
\usage{
enrichMsigDB(
  gene_list,
  from_type = "SYMBOL",
  orgdb = org.Mm.eg.db::org.Mm.eg.db,
  msigdb_category = "H",
  species = "Mus musculus",
  background = NULL,
  col_genetype = "GENETYPE",
  feature_type = "protein-coding",
  ...
)
}
\arguments{
\item{gene_list}{a character vector denoting valid gene names.}

\item{from_type}{a string, default SYMBOL, denoting the type of input gene_list (e.g SYMBOL, ENTREZID).}

\item{orgdb}{an object of the class OrgDB. Default org.Mm.eg.db::org.Mm.eg.db.}

\item{msigdb_category}{a character string, default "H", denoting one of the nine catagories of geneList given in the msigdb. Possible values are: "H","C1","C2","C3","C4","C5","C6","C7", and "C8".}

\item{species}{a character string denoting name of the species. Default "Mus musculus". Possible values can be found via the function \code{\link[msigdbr:msigdbr_species]{msigdbr::msigdbr_species()}}.}

\item{background}{\code{NULL} (default) or a character vector. If \code{NULL}, all features filtered by \code{feature_type} will be selected as background.
In case of character vector, values will be used as it is for background set. In case of user provided background set it is assumed that id type is same as queried \code{genes_list} (i.e. \code{from_type}).
NOTE: In both scenarios values obtained from MSigDB catagory database will always remain as a part of background set.}

\item{col_genetype}{A string denoting a column name to filter \code{feature_type}.Valid only if the argument \code{background} is \code{NULL}.}

\item{feature_type}{A string denoting a value to filter the column mentioned by the argument \code{col_genetype}.Valid only if the argument \code{background} is \code{NULL}.}

\item{...}{Other parameters to pass \code{\link[clusterProfiler:enricher]{clusterProfiler::enricher()}}.}
}
\value{
Enrichment results.
}
\description{
Perform over representation analysis (ORA) against Molecular Signature Database (MsigDB) .
}
\examples{
\dontrun{

}
}
