% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicewiz_wrappers.R
\name{get_ASE_data_matrix}
\alias{get_ASE_data_matrix}
\title{Get a data matrix (PSI, logit, z-score) for a given ASE}
\usage{
get_ASE_data_matrix(
  se,
  event_names,
  samples,
  summarise_groups = TRUE,
  summarise_groups_by = "mean",
  method = "PSI",
  column_condition = "condition"
)
}
\arguments{
\item{se}{an object of class NxtSE.}

\item{event_names}{a character vector denoting valid ASE names to plot in the heatmap.}

\item{samples}{a character vector denoting valid sample names to plot in the heatmap.}

\item{summarise_groups}{logical, denoting whether to summaries groups (replicates mostly), default \code{TRUE}.}

\item{summarise_groups_by}{a character string denoting a method to summaries groups. Can be one of the \code{mean} or \code{median}.}

\item{method}{a character string denoting a method for data value. Can be one of the "PSI", "logit" or "Z-Score".}

\item{column_condition}{a character string denoting a condition column from \code{se} to select samples. Same column will also be used to group and summaries samples.}
}
\value{
a data matrix
}
\description{
Get a data matrix (PSI, logit, z-score) for a given ASE
}
\examples{
se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
SpliceWiz::colData(se)$treatment <- rep(c("A", "B"), each = 3)
SpliceWiz::colData(se)$replicate <- rep(c("P","Q","R"), 2)
res <- run_ase_diff_analysis(x = se, test_factor = "treatment", test_nom = "A" ,test_denom = "B",  IRmode ="annotated",  cutoff_lfc = 0.6, cutoff_padj = 1, regul_based_upon = 2)
event_names = get_ASEsets_by_regulation(x = res, sample_comparisons = "A_VS_B", regul = "all") \%>\% unlist()
get_ASE_data_matrix(se, event_names , samples = c("A", "B"), column_condition ="treatment")
get_ASE_data_matrix(se, event_names , samples = c("P", "Q","R"), column_condition ="replicate")

}
