% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{get_diff_gene_count_barplot}
\alias{get_diff_gene_count_barplot}
\title{Generate a barplot of DEG counts.}
\usage{
get_diff_gene_count_barplot(
  x,
  col_up = "#a40000",
  col_down = "#16317d",
  font_size = 12
)
}
\arguments{
\item{x}{an object of class parcutils}

\item{col_up}{a character string, default #a40000, denoting valid a color name for "Up" regulated genes.}

\item{col_down}{a character string, default #16317d, denoting valid a color name "Down" regulated genes.}

\item{font_size}{a numeric, default 12, denoting font size in the plot.}
}
\value{
a ggplot
}
\description{
Generate a barplot of DEG counts.
}
\examples{
count_file <- system.file("extdata","toy_counts.txt" , package = "parcutils")
count_data <- readr::read_delim(count_file, delim = "\t")

sample_info <- count_data \%>\% colnames() \%>\% .[-1]  \%>\%
 tibble::tibble(samples = . , groups = rep(c("control" ,"treatment1" , "treatment2"), each = 3) )


res <- run_deseq_analysis(counts = count_data ,
                         sample_info = sample_info,
                         column_geneid = "gene_id" ,
                         group_numerator = c("treatment1", "treatment2") ,
                         group_denominator = c("control"))

get_diff_gene_count_barplot(res)
}
