% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{get_gene_expression_box_plot}
\alias{get_gene_expression_box_plot}
\title{Generate a box plot}
\usage{
get_gene_expression_box_plot(
  x,
  samples = NULL,
  genes = NULL,
  group_replicates = F,
  convert_log2 = T
)
}
\arguments{
\item{x}{an abject of class "parcutils". This is an output of the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}.}

\item{samples}{a character vector denoting samples to plot in boxplot, default \code{NULL}. If set to NULL all samples are accounted.}

\item{genes}{a character vector denoting genes to consider for boxplot, default \code{NULL}. If set to NULL all genes are accounted.}

\item{group_replicates}{logical, default \code{FALSE}, whether to group replicates in individual plots.}

\item{convert_log2}{logical, default \code{TRUE}, whether to plot log2(noralised expression + 1) or not.}
}
\value{
an object of class ggplot2
}
\description{
Generate a box plot
}
\examples{
\dontrun{
 // TO DO
}


}
