% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{get_gene_expression_heatmap}
\alias{get_gene_expression_heatmap}
\title{Given a set of genes and sample names create a heatmap of gene expression value or z-score.}
\usage{
get_gene_expression_heatmap(
  x,
  samples,
  genes,
  repair_genes = FALSE,
  convert_log2 = FALSE,
  color_default = TRUE,
  col = NULL,
  convert_zscore = TRUE,
  summarise_replicates = TRUE,
  summarise_method = "median",
  show_row_names = FALSE,
  cluster_rows = TRUE,
  show_row_dend = TRUE,
  row_names_font_size = 10,
  show_column_names = TRUE,
  cluster_columns = TRUE,
  show_heatmap_legend = TRUE,
  ...
)
}
\arguments{
\item{x}{an abject of class "parcutils". This is an output of the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}.}

\item{samples}{a character vector denoting sample names to show in the heatmap.}

\item{repair_genes}{logical, default FALSE, indicating whether to repair gene names or not. See details.}

\item{convert_log2}{logical, default FALSE, indicating whether to convert gene expression values in log2 or not.}

\item{color_default}{logical, default TRUE, indicating whether to use default heatmap colors.}

\item{col}{an output of \code{\link[circlize:colorRamp2]{circlize::colorRamp2()}}, default NULL.}

\item{convert_zscore}{logical, default TRUE, indicating whether to convert gene expression values in to the z-score or not. see details.}

\item{summarise_replicates}{logical, default TRUE, indicating whether to summarise values for each gene across replicates.}

\item{summarise_method}{a character string, either mean or median.}

\item{show_row_names}{logical, default FALSE, indicating whether to show row names in the heatmap or not.
Internally this argument is passed to the function \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.}

\item{cluster_rows}{logical, default TRUE, indicating whether to cluster rows in the heatmap or not.
Internally this argument is passed to the function \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.}

\item{show_row_dend}{logical, default TRUE, indicating whether to show dendrogram in the heatmap or not.
Internally this argument is passed to the function \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.}

\item{row_names_font_size}{a numeric value, default 10, indicating size of row names in the heatmap.}

\item{show_column_names}{logical, default TRUE, indicating whether to show column names in the heatmap or not.
Internally this argument is passed to the function \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.}

\item{cluster_columns}{logical, default TRUE, indicating whether to cluster columns in the heatmap or not.
Internally this argument is passed to the function \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.}

\item{show_heatmap_legend}{logical, default TRUE, indicating whether to show heatmap legend or not.}

\item{...}{Other parameters to be passes to \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.}

\item{genesa}{character vector denoting genes to show in the heatmap.}
}
\value{
an output of the function \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.
}
\description{
Heatmap is a common tool to show gene expression pattern across samples in RNAseq experiment.
While doing data exploration, it is a common practice to generate several heatmaps to identify interesting
patterns of gene expression across sample. Most heatmap generating tools require data in a tabular form.
However, to prepare such a tabular form requires significant data wrangling such as subsetting
genes (rows) and samples (columns) specifically when RNASeq studies studies consisting of several samples and
replicates.

This function cut downs several steps of data wrangling to create a heatmap of gene expression. Given an object of
'parcutils', names of samples, genes to show in heatmap and several other arguments it creates a heatmap.
The output of the function is an output of the function \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}} which then can be used for
the other functions of the \link{ComplexHeatmap} pacakge.
}
\details{
\code{repair_genes} Internally gene names are stored as a "gene_id:gene_symbol" format. For example, "ENSG00000187634:SAMD11".
When \code{repair_genes} is set to \code{TRUE} the string corresponding to gene_id followed by ":" will be removed. This is useful when gene names
are revealed in the heatmap.
\code{convert_zscore}. When set to \code{TRUE} values for each gene is converted into z-score. z-score is calculated by baser r function \code{\link[base:scale]{base::scale()}}
with all default parameters.
}
\examples{
\dontrun{
// TO DO.
}

}
