% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{get_genes_by_regulation}
\alias{get_genes_by_regulation}
\title{Get genes based on their differential regulation}
\usage{
get_genes_by_regulation(x, sample_comparison, regulation = "both")
}
\arguments{
\item{x}{an abject of class "parcutils". This is an output of the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}.}

\item{sample_comparison}{a character string denoting a sample comparison for which genes to be obtained.}

\item{regulation}{a character string, default \code{both}. Values can be one of the \code{up}, \code{down}, \code{both}, \code{other}, \code{all}.
\itemize{
\item up : returns all up regulated genes.
\item down : returns all down regulated genes.
\item both : returns all up and down regulated genes.
\item other : returns genes other than up and down regulated genes.
\item all : returns all genes.
}}
}
\value{
a named vector.
}
\description{
Get genes based on their differential regulation
}
\details{
For a given differential comparison this function returns \code{up}, \code{down}, \code{both}, \code{other} and \code{all} genes.
}
\examples{
\dontrun{
// TO DO
}

}
