% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{get_normalised_expression_matrix}
\alias{get_normalised_expression_matrix}
\title{Prepare a matrix of normalised gene expression values}
\usage{
get_normalised_expression_matrix(
  x,
  sample_comparisons,
  genes,
  summarise_replicates = FALSE,
  summarise_method = "median"
)
}
\arguments{
\item{x}{an abject of class "parcutils". This is an output of the function \link[parcutils]{run_deseq_analysis}.}

\item{sample_comparisons}{a character vector denoting sample comparisons for which fold change values to be derived.}

\item{genes}{a character vector denoting gene names for which fold change values to be derived.}

\item{summarise_replicates}{logical, default FALSE, indicating whether gene expression values summarised by mean or median between replicates.}

\item{summarise_method}{a character string either "mean" or "median" by which normalised gene expression values will be summarised between replicates.}
}
\value{
a tbl.
}
\description{
This function returns a dataframe having first column gene names and subsequent columns are
normalised gene expression values for the comparisons passed through sample_comparisons.
}
\examples{
\dontrun
{
// TO DO.
}
}
