% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_bioinfo.R
\name{get_star_align_log_summary}
\alias{get_star_align_log_summary}
\title{Get alignment summary of STAR aligner}
\usage{
get_star_align_log_summary(log_file)
}
\arguments{
\item{log_file}{a string denoting a path to *Log.final.out file generated by STAR aligner.}
}
\value{
a tbl.
}
\description{
Get alignment summary of STAR aligner
}
