% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_related.R
\name{get_star_align_log_summary_plot}
\alias{get_star_align_log_summary_plot}
\title{Visualise alignemnt stats in a bar plot}
\usage{
get_star_align_log_summary_plot(
  x,
  col_total_reads = "#007e2f",
  col_mapped_reads = "#ffcd12"
)
}
\arguments{
\item{x}{a named character vector denoting paths to *Log.final.out file generated by STAR aligner.
Names will be used to label Y axis in the plot. If not named, file names will be used to label.}

\item{col_total_reads}{a character string denoting bar color total reads. Default "#007e2f".}

\item{col_mapped_reads}{a character string denoting bar color total reads. Default "#ffcd12".}
}
\value{
a bar plot.
}
\description{
Visualise alignemnt stats in a bar plot
}
\examples{

file_dir <- system.file("extdata",package = "parcutils")

file_paths <- fs::dir_ls(path = file_dir,
glob = "*Log.final.out" , recurse = T, type = "file") \%>\% as.character()

get_star_align_log_summary_plot(x  = file_paths)
get_star_align_log_summary_plot(x  = file_paths, col_total_reads = "red",
col_mapped_reads = "blue")


}
