% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_related.R
\name{make_enriched_heatmap_list}
\alias{make_enriched_heatmap_list}
\title{Create a HeatmapList from parcutils_chip.}
\usage{
make_enriched_heatmap_list(
  x,
  row_order = NULL,
  cluster_targets_by_rpm = TRUE,
  cluster_by = NULL,
  n_clust = 2,
  cluster_rows = FALSE,
  heatmap_columns = NULL,
  heatmap_column_title = NULL,
  heatmap_pos_line = FALSE,
  heatmap_pos_line_gp = grid::gpar(lty = 2),
  heatmap_colors = circlize::colorRamp2(breaks = c(0, 0.5, 1), colors = c("#fee6ce",
    "#e6550d", "#E31A1C")),
  heatmap_top_annotations = NULL,
  heatmap_axis_name = c("-3KB", "Summit", "+3KB"),
  heatmap_axis_name_rot = 0,
  heatmap_axis_name_gp = gpar(fontsize = 10),
  heatmap_border = TRUE
)
}
\arguments{
\item{x}{an object of the class parcutils_chip.}

\item{row_order}{internally passed to \code{row_order} argument of \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.}

\item{cluster_targets_by_rpm}{logical, denoting weather to cluster targets by RPM value.}

\item{cluster_by}{NULL (default) or character vector denoting RPM columns to be used for clustering.}

\item{n_clust}{a numeric, default 2, denoting number of clusters for kmeans clustering.}

\item{cluster_rows}{logical, default FALSE, denoting weather to cluster rows using default clustering. This must be always false as kmeans is implimented.}

\item{heatmap_columns}{NULL (default) or character vector denoting columns to be displayed in the heatmap.}

\item{heatmap_column_title}{NULL (default) or character vector denoting title for each heatmap column.}

\item{heatmap_pos_line}{logical, default FALSE, internally passed to \code{pos_line} argument of \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.}

\item{heatmap_pos_line_gp}{logical, default FALSE, internally passed to \code{pos_line_gp} argument of \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.}

\item{heatmap_colors}{internally passed to \code{col} argument of \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.
Default \code{circlize::colorRamp2(breaks = c(0,0.5,1),colors = c('#fee6ce','#e6550d','#E31A1C'))}.}

\item{heatmap_top_annotations}{internally passed to \code{top_annotations} argument of \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.}

\item{heatmap_axis_name}{internally passed to \code{axis_name} argument of \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.}

\item{heatmap_axis_name_rot}{internally passed to \code{axis_name_rot} argument of \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.}

\item{heatmap_axis_name_gp}{internally passed to \code{axis_name_gp} argument of \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.}

\item{heatmap_border}{internally passed to \code{border} argument of \code{\link[EnrichedHeatmap:EnrichedHeatmap]{EnrichedHeatmap::EnrichedHeatmap()}}.}
}
\value{
a HeatmapList.
}
\description{
Create a HeatmapList from parcutils_chip.
}
\examples{
\dontrun{
}

}
