% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_bioinfo.R
\name{normalise_counts}
\alias{normalise_counts}
\title{Convert raw read counts in to normalised values.}
\usage{
normalise_counts(x, .vars = NULL, method = "TPM")
}
\arguments{
\item{x}{A dataframe of raw counts along with mandatory columns which are \code{GeneID, Chr, Start, End, Strand, Length}.}

\item{.vars}{A character vector containing columns from \code{x}. Normalization will be performed only on these columns. If NULL (default) all columns normlaisation will be
performed on all columns.}

\item{method}{A character string, default TPM. Choices are one of TPM, RPKM.}
}
\value{
A dataframe with all mandatory columns along with columns mentioned in \code{.vars}. Remaining columns from \code{x} will be dropped.
}
\description{
Reads Per Kilo Base Per Million (RPKM) and Tags Per Million (TPM) are two most waidely used normalisation methods in RNA-seq experiments.
This function convert raw read counts either into RPKM or TPM values.
}
\details{
implemntaion of RPKM and TPM can be seen in functions \code{\link[=get_tpm]{get_tpm()}} and \code{\link[=get_rpkm]{get_rpkm()}} respectively.
}
\examples{
set.seed(123)
tt <- tibble::tibble(gene_id = c(paste("Gene_",1:5,sep = "")),
                     chr = "Chr1",
                     start = sample(1:100, 5),
                     end = sample(100:200,5),
                     strand = sample(c("+" ,"-"), 5, replace = T),
                     length = (end - start )+ 1 )

tt \%<>\% dplyr::mutate(sample_1 = sample(c(1:100),5)*10 ,
                      sample_2 = sample(c(1:100),5)*10 ,
                      sample_3 = sample(c(1:100),5)*100,
                      sample_4 = sample(c(1:100),5)*100 )

normalise_counts(x = tt ,method = "RPKM")
normalise_counts(x = tt, .vars = c("sample_1","sample_2") ,method = "TPM")
normalise_counts(x = tt, .vars = c("sample_1","sample_2") ,method = "RPKM")
}
\seealso{
\code{\link[=get_tpm]{get_tpm()}}  \code{\link[=get_rpkm]{get_rpkm()}}
}
