% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_related.R
\docType{class}
\name{parcutils_chip-class}
\alias{parcutils_chip-class}
\alias{parcutils_chip}
\title{Set class parcutils_chip}
\value{
an object of class parcutils_chip
}
\description{
Set class parcutils_chip
}
\examples{
\dontrun{
<<<<<<< HEAD
#// TO DO
=======
// TO DO
>>>>>>> 5c15ed662e9e47e5e08c1033089c800845d9f9d2
}
}
\keyword{internal}
