% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{piarwise_upset}
\alias{piarwise_upset}
\title{Generate upset plots for DEG between a comparison.}
\usage{
piarwise_upset(
  x,
  sample_comparison,
  color_up = "#b30000",
  color_down = "#006d2c",
  ...
)
}
\arguments{
\item{x}{an abject of class "parcutils". This is an output of the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}.x}

\item{sample_comparison}{a character vector of length 2 denoting  sample comparisons between upset plot to be generated.}

\item{color_up}{a character string denoting a valid color code for bars in upset plot for up regulated genes.}

\item{color_down}{a character string denoting a valid color code for bars in upset plot for down regulated genes.}

\item{...}{other arguments to be passed to \code{\link[UpSetR:upset]{UpSetR::upset()}}.}
}
\value{
an object of \code{\link[UpSetR:upset]{UpSetR::upset()}}
}
\description{
Generate upset plots for DEG between a comparison.
}
\examples{
\dontrun{

// TO DO.
}

}
\keyword{internal}
