% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_related.R
\name{plot_deg_upsets}
\alias{plot_deg_upsets}
\title{Generate upset plots for differently expressed genes between comparisons.}
\usage{
plot_deg_upsets(
  x,
  sample_comparisons,
  color_up = "#b30000",
  color_down = "#006d2c"
)
}
\arguments{
\item{x}{x an abject of class "parcutils". This is an output of the function \code{\link[=run_deseq_analysis]{run_deseq_analysis()}}.}

\item{sample_comparisons}{a character vector denoting  sample comparisons between upset plot to be generated.}

\item{color_up}{a character string denoting a valid color code for bars in upset plot for up regulated genes.}

\item{color_down}{a character string denoting a valid color code for bars in upset plot for down regulated genes.}
}
\value{
an object of named list where each element is a list of two - 1)  an upset plots  \code{\link[UpSetR:upset]{UpSetR::upset()}} and their intersections in form of tibble.
}
\description{
For a given set of DEG comparisons, the functions returns \code{\link[UpSetR:upset]{UpSetR::upset()}} plots for up and down genes between any 2 comparisons.
For each upset plot generated function also returns interaction between gene sets in form of dataframe.
}
\examples{
\dontrun{

// TO DO.
}

}
